### 2021-11-23 #3945 @whedon
[PRE REVIEW]: Castor: A C++ library to code "à la matlab"
**Submitting author:** @matthieuaussal (<a href="http://orcid.org/0000-0002-2812-7578">Matthieu Aussal</a>)
**Repository:** <a href="https://github.com/leprojetcastor/castor" target ="_blank">https://github.com/leprojetcastor/castor</a>
**Version:** v1.0.0
