### 2021-11-23 #3945 @jbytecode
Dear @matthieuaussal - maybe you can use [https://whedon.theoj.org/](https://whedon.theoj.org/) to debug your md file and make it successfully compiled to pdf.
