### 2021-11-23 #3945 @xuanxu
The PDF is not compiling due to unscaped characters. [This change](https://github.com/leprojetcastor/castor/pull/1) should fix it.
