### 2021-11-29 #3945 @matthieuaussal
Dear Editor,

We can suggest @agramfort, wich is a colleague from INRIA.

Otherwise, from your list, we check people both in Matlab and C++ preferred langage, with a numerical background  :

Applied math :
@junliu2050
@rosoport
@faroukm
@pitsianis
@rosoport

Others (signal processing, mechanics, etc.):
@greatv
@hasantahir
@florian-rosenthal
@somefunagba
@declanol
@ononak
@WalBytes
@AdeBytes

Is it ok for you?

Best,

Matthieu

> Le 28 nov. 2021 à 13:16, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> Dear @matthieuaussal <https://github.com/matthieuaussal>
> Do you have any suggestions for potential reviewers? You can also use the list here <https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822>.
> Please mention the GitHub users without using the @ character.
>
> Thank you in advance.
>
