### 2021-11-30 #3945 @jbytecode
👋👋👋 @junliu2050 @rosoport 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "Castor: A C++ library to code "à la matlab". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3945).
