### 2021-11-30 #3945 @mkitti
I'm not clear if there is a rendering issue or not, but the figures appear to lack captions.
