### 2021-11-30 #3945 @jbytecode
@mkitti when the review starts, you can address this issue again so the authors corrects the paper so that the figures are linked to the caption.
