### 2021-11-30 #3945 @junliu2050
Hi, thanks for inviting me. I guess I can try to provide some feedback
based on my background in scientific computing.
Will the review process be anonymous or not? I think I like the idea of
JOSS, but never had any experience with it yet.

On Tue, Nov 30, 2021 at 2:30 AM Mehmet Hakan Satman <
***@***.***> wrote:

> 👋👋👋 @junliu2050 <https://github.com/junliu2050> @rosoport
> <https://github.com/rosoport> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "Castor: A C++ library to code
> "à la matlab". You can find more information at the top of this Github
> issue (#3945 <https://github.com/openjournals/joss-reviews/issues/3945>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
