### 2021-12-01 #3965 @whedon
[REVIEW]: Castor: A C++ library to code "à la matlab"
**Submitting author:** @matthieuaussal (<a href="http://orcid.org/0000-0002-2812-7578">Matthieu Aussal</a>)
**Repository:** https://github.com/leprojetcastor/castor
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
