### 2021-12-01 #3965 @jbytecode
Dear @matthieuaussal

I have sent a pull request. It includes the missing DOIs that whedon suggests. Please apply the patch if you are agreed with me.

thank you in advance.
