### 2021-12-01 #3965 @matthieuaussal
> Le 1 déc. 2021 à 13:09, whedon ***@***.***> a écrit :
>
>
> I'm sorry human, I don't understand that. You can see what commands I support by typing:
>
> @whedon commands
