### 2021-12-01 #3945 @matthieuaussal
@mkitti we forget to define captions, sorry for that. It’s now done!

> Le 30 nov. 2021 à 21:21, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> @mkitti <https://github.com/mkitti> when the review starts, you can address this issue again so the authors corrects the paper so that the figures are linked to the caption.
>
