### 2021-12-01 #3945 @matthieuaussal
So we do not make any changes until the review report has been sent? (it’s my first submission in joss)

> Le 1 déc. 2021 à 13:21, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> @matthieuaussal <https://github.com/matthieuaussal> this is the pre-review issue, all issues belong to the paper review is discussed and reported in the review issue.
>
