### 2021-12-01 #3945 @junliu2050
Hi, after carefully reading through the paper, I feel I may not be very
suitable for reviewing such a library since I do not use or develop a
similar package by myself.
Hence you may assign Rodrigo Portugal as 2nd reviewer. Overall it is an
interesting package, but the examples and comparison with similar packages
can be improved.

On Wed, Dec 1, 2021 at 10:56 AM Mehmet Hakan Satman <
***@***.***> wrote:

> @rosoport <https://github.com/rosoport> thank you for your response,
> however, we have already assigned two reviewers to this submission. Hope we
> work together in the future.
>
