### 2021-12-09 #3965 @mkitti
The authors present, Castor, a header-only C++ template library that emulates the high-level semantics of Matlab, a proprietary commerical product licensed by Mathworks, Inc (Natick, MA). This includes several classes to represent dense, sparse, and hierarchical matrices. They combine this with operators and functions to manipulate these matrices and perform linear algebra operations. This combination of the header library with matrix manipulation and linear algebra appears to be the primary contribution. Additionally, the authors have added a graphics subsystem based on VTK to help visualize the results of the matrix operations.
