### 2021-12-09 #3965 @mkitti
# Remarks on Summary, Statement of Need, and State of the Field

An important task for the paper and in the checklist is to establish who the audience is. As written, the paper seem mainly oriented at the MATLAB user who perhaps would like to port their code to C++. This would perhaps be motivated by seeking any of restrictions of the proprietary licensing of MATLAB.

There are several alternatives to MATLAB based licensing including using Octave, NumPy, or Julia. There are all available under open source licenses, provide matrix manipulation libraries, and provide methods for visualizing the results. The summary mentions "Python", however, I think it would be worthwhile to specifically mention NumPy since that is the package in Python necessary for general matrix manipulation and linear algebra.

The critique of MATLAB, Python, and Julia made by the authors is that these are interpreted languages. In the case of MATLAB, there is increasing of use a just-in-time (JIT) compiler via the MATLAB Execution Engine in R2015b. Python does have some compiler frameworks available via Numba (via LLVM) or PyPy (JIT).

Julia is misidentified here as an interpreted language since the normal mode of executing Julia is to compile it to native code via LLVM, at first execution without a discrete compilation step. While a Julia interpreter does exist, it is mostly limited to use during debugging. Furthermore, compilation is an important factor in the use of Julia since it contributes considerable [compilation latency](https://julialang.org/blog/2020/08/invalidations/) during initial execution.

This line of critique also ignores the possibility of using a C++ interpreter in conjunction with this library such as [Cling](https://root.cern/cling/), a part of ROOT.. This would permit a more complete MATLAB-like experience particularly in offering a read-evaluate-print-loop (REPL) interface. Another her possibility is Cxx.jl, a Julia package that includes a C++ interpreter.
https://compiler-research.org/assets/presentations/O_Schulz-Julia_ROOT.pdf
https://compiler-research.org/assets/presentations/K_Fischer_Cxx_jl.pdf

Another potential audience for this work is the C++ developer who would like perform tasks similar to MATLAB. In this space, we need to consider other C++ matrix manipulation and linear algebra libraries such as [Eigen](https://eigen.tuxfamily.org/index.php?title=Main_Page). A direct comparison to Eigen in this paper would be helpful since it would help a C++ programmer decide between the two C++ template libraries.

Castor is an open-source C++ matrix manipulation library that implements a MATLAB-like API by using similar function names as MATLAB. In total, this perhaps could differentiate this package from other matrix manipulation languages.

To summarize, there are several potential audiences for this work:
1. The MATLAB user who would like to port code to C++
2. The open source developer would like to perform matrix-manipulation through ahead-of-time compilation.
3. The C++ developer would like to matrix-manipulation similar in fashion to MATLAB

The summary and statement of need primarily focuses on audience #1, the MATLAB-user. The authors argue that an IDE greatly  simplifies the use of C++. However, I believe there are many other important considerations such as memory management, garbage collection, and programming background which would need to be considered for a MATLAB-user to start using C++. An IDE alone does not eliminate the need to learn a new language and memory management concepts. This potential audience  is likely limited and size and scope.

The authors should address audiences #2 and #3 directly. This library is immediately usable to someone who already has a C++ development environment, especially if they are also a MATLAB user or work with MATLAB users. For this reason, I think a comparison with other C++ matrix manipulation libraries such as Eigen is needed here. While Eigen has a distinct API from MATLAB, the APIs of NumPy and Julia are also distinct. Yet these libraries have existing and expanding userbases.

The authors should clarify the above points in the Summary and Statement of Need section. They should also add a clearly identifiable "State of the Field" section where they more clearly discuss the relationship of the software to MATLAB, other open source matrix manipulation libraries, and other C++ matrix manipulation libraries such as Eigen.

# Revisions Checklist

- [ ] Consider mentioning the included VTK-based graphics system
- [ ] Clarify how IDEs provide flexibility of executing compiled code with flexibility similar to interpreted language, or remove.
- [ ] Clarify that Julia is not an interpreted language, and perhaps differentiate from Julia by emphasizing ahead-of-time compilation

- [ ] Address what this library adds beyond other existing open-source libraries for matrix-manipulation
- [ ] Consider adding potential use in a C++ interpreter such as Cling

- [ ] Add this section
- [ ] Compare Castor to the C++ matrix manipulation library, Eigen, and how the goals here may differ from that project
- [ ] Compare to NumPy, as a popular open-source matrix manipulation library, with some compilation capability via Numba
- [ ] Compare Castor to the use of Julia, which is a compiled matrix manipulation language
