### 2021-12-10 #3965 @matthieuaussal
@mkitti Many thanks for your review. We push an update which takes into account your remarks:

 + Summary :
 - We add a mention to BLAS and VTK
 - We remove IDE
 - We remove Julia

+ Statement of need
 - We precise that no existing libraries are same to matlab semantics for matrix manipulations, algebra and visu
 - We add a word to notebook using Cling

+ Statement of the field
We are not enthusiastic about the idea of a comparison with the tools mentioned, because that is not our goal, nor our motivation. Indeed, we simply sought to create a library as close as possible to Matlab semantics, to facilitate the work of developers who would like to find their Matlab reflexes in C ++ (it's exactly our case). We did this job precisely because none of the libraries mentioned met our needs:
  - Semantic (almost) identical to Matlab ('transpose (A)' and not 'A.transpose ()')
  - C ++ template language
  - Portable (no interpreter)
  - Dense, sparse, hierarchical, algebra and visual formats
What is your opinion?

Best regards

Matthieu, Laurent and Marc
