### 2021-12-10 #3965 @mkitti
Thank you for the revisions.

The **summary** is much clearer now, and it fully encompasses the scope of the software.

For the **statement of the need**, discuss in general why the MATLAB semantics you do implement and why those are important and perhaps sufficient because certainly you do not implement all of MATLAB's semantics. For example, as far as I can tell, you do not have the [backslash operator](https://blogs.mathworks.com/cleve/2013/08/19/backslash/)  which Cleve Moler, the creator of MATLAB, states

> The backslash operator has come to represent both the matrix origins and the usability of MATLAB.

As far as I know, `\` is not a valid operator in C++. Perhaps you could implement a function [`mldivide`](https://www.mathworks.com/help/matlab/ref/mldivide.html). That is far from trivial (see the flowchart), but it would really emulate much of the MATLAB experience rather than just being BLAS wrapper that looks similar to MATLAB.

# State of the Field

I just want to emphasize this is a **required** section that is missing. Part of the checklists asks ["Do the authors describe how this software compares to other commonly-used packages?"](https://joss.readthedocs.io/en/latest/review_checklist.html?highlight=field#software-paper). Some kind of comparison is needed here, and perhaps that comparison would involve explaining why the semantics of these packages differ from MATLAB semantics.

For the `transpose` example you gave, I can do this with NumPy:
```python
In [1]: import numpy as np

In [2]: from numpy import transpose

In [3]: A = np.random.randint(1,10, (5,5))

In [4]: transpose(A)
Out[4]:
array([[6, 4, 4, 2, 1],
       [4, 8, 5, 3, 4],
       [4, 9, 6, 8, 4],
       [4, 6, 3, 1, 3],
       [2, 2, 7, 4, 4]])
```

I would argue that the typical semantics for that operation in MATLAB would be `A.'` which is not valid syntax in Castor/C++ or NumPy/Python. There are other comparisons such as in this [cheatsheet](https://numpy.org/doc/stable/user/numpy-for-matlab-users.html).

Similarly, in Julia there is also a `transpose` and a tick operator:

```julia
julia> transpose(A)
5×5 transpose(::Matrix{Int64}) with eltype Int64:
 5  7  4  9  9
 1  7  9  3  4
 2  7  4  4  5
 5  1  7  3  3
 5  5  9  2  2

julia> A'
5×5 adjoint(::Matrix{Int64}) with eltype Int64:
 5  7  4  9  9
 1  7  9  3  4
 2  7  4  4  5
 5  1  7  3  3
 5  5  9  2  2
```

The above are actually lazy operations, but work similarly to their eager implementations. We can also show that in Julia, there are also similar eager operations which are compiled.

<summary>Native code of Julia's `LinearAlgebra.transpose_f!`</summary>

```julia
julia> @code_native LinearAlgebra.transpose_f!(transpose, B, A)
        .text
; ┌ @ transpose.jl:95 within `transpose_f!`
        pushq   %rbp
        movq    %rsp, %rbp
        pushq   %r15
        pushq   %r14
        pushq   %r13
        pushq   %r12
        pushq   %rsi
        pushq   %rdi
        pushq   %rbx
        subq    $168, %rsp
        movq    %rdx, %rbx
        vxorps  %xmm0, %xmm0, %xmm0
        vmovaps %xmm0, -128(%rbp)
        movq    $0, -112(%rbp)
        movq    %rbx, -168(%rbp)
        movl    $jl_get_pgcstack, %eax
        callq   *%rax
        movq    $4, -128(%rbp)
        movq    (%rax), %rcx
        movq    %rcx, -120(%rbp)
        leaq    -128(%rbp), %rcx
        movq    %rcx, (%rax)
        movq    8(%rbx), %rcx
        movq    16(%rbx), %rdx
; │ @ transpose.jl:96 within `transpose_f!`
; │┌ @ abstractarray.jl:95 within `axes`
; ││┌ @ array.jl:152 within `size`
        movq    24(%rdx), %r8
        movq    32(%rdx), %r9
; │└└
; │ @ transpose.jl:97 within `transpose_f!`
; │┌ @ abstractarray.jl:74 within `axes` @ abstractarray.jl:95
; ││┌ @ array.jl:152 within `size`
        movq    24(%rcx), %r12
; │└└
; │┌ @ range.jl:1043 within `==`
        cmpq    $1, %r12
        je      L116
        testq   %r12, %r12
        jne     L124
; ││┌ @ range.jl:609 within `isempty`
; │││┌ @ operators.jl:378 within `>`
; ││││┌ @ int.jl:83 within `<`
        testq   %r9, %r9
; │└└└└
        je      L129
        jmp     L153
; │┌ @ range.jl:1044 within `==`
; ││┌ @ range.jl:1057 within `_has_length_one`
; │││┌ @ promotion.jl:468 within `==`
L116:
        cmpq    $1, %r9
; │└└└
        je      L129
        jmp     L153
; │┌ @ range.jl:1045 within `==` @ promotion.jl:468
L124:
        cmpq    %r9, %r12
; │└
        jne     L153
; │┌ @ abstractarray.jl:74 within `axes` @ abstractarray.jl:95
; ││┌ @ array.jl:152 within `size`
L129:
        movq    32(%rcx), %rbx
; │└└
; │┌ @ range.jl:1043 within `==`
        cmpq    $1, %rbx
        je      L230
        testq   %rbx, %rbx
        jne     L752
; ││┌ @ range.jl:609 within `isempty`
; │││┌ @ operators.jl:378 within `>`
; ││││┌ @ int.jl:83 within `<`
        testq   %r8, %r8
; │└└└└
        je      L236
L153:
        movq    $316687360, -96(%rbp)           # imm = 0x12E04400
        movabsq $jl_apply_generic, %rdi
        leaq    -96(%rbp), %rsi
        movl    $291737952, %ecx                # imm = 0x11639160
        movq    %rsi, %rdx
        movl    $1, %r8d
        callq   *%rdi
        movq    %rax, -112(%rbp)
        movq    %rax, -96(%rbp)
        movl    $351419696, %ecx                # imm = 0x14F23D30
        movq    %rsi, %rdx
        movl    $1, %r8d
        callq   *%rdi
        movabsq $jl_throw, %rdx
        movq    %rax, %rcx
        callq   *%rdx
; │┌ @ range.jl:1044 within `==`
; ││┌ @ range.jl:1057 within `_has_length_one`
; │││┌ @ promotion.jl:468 within `==`
L230:
        cmpq    $1, %r8
; │└└└
        jne     L153
; │ @ transpose.jl within `transpose_f!`
L236:
        movq    %rax, -136(%rbp)
; │ @ transpose.jl:100 within `transpose_f!`
; │┌ @ int.jl:88 within `*`
        movq    %r9, %rax
        imulq   %r8, %rax
; │└
; │┌ @ int.jl:477 within `<=`
        cmpq    $256, %rax                      # imm = 0x100
        movq    %rcx, -72(%rbp)
; │└
        jle     L289
; │ @ transpose.jl:109 within `transpose_f!`
        vxorps  %xmm0, %xmm0, %xmm0
        vmovups %xmm0, 32(%rsp)
        movabsq $"transposeblock!", %rax
        callq   *%rax
        jmp     L711
; │ @ transpose.jl within `transpose_f!`
L289:
        movq    %r12, -64(%rbp)
; │ @ transpose.jl:102 within `transpose_f!`
; │┌ @ range.jl:833 within `iterate`
; ││┌ @ range.jl:609 within `isempty`
; │││┌ @ operators.jl:378 within `>`
; ││││┌ @ int.jl:83 within `<`
        testq   %r9, %r9
; │└└└└
        je      L711
; │ @ transpose.jl within `transpose_f!`
        testq   %r8, %r8
; │ @ transpose.jl:103 within `transpose_f!`
        je      L711
; │ @ transpose.jl within `transpose_f!`
        movq    (%rdx), %rcx
        movq    -72(%rbp), %rax
        movq    (%rax), %r11
; │ @ transpose.jl:104 within `transpose_f!`
        leaq    96(%r11), %rbx
        leaq    96(%rcx), %rsi
        leaq    (,%r8,8), %rax
        movq    %rax, -144(%rbp)
        movq    %r8, %r10
        andq    $-16, %r10
        movq    -64(%rbp), %rax
        leaq    (,%rax,8), %r13
        movq    %rcx, -88(%rbp)
        leaq    -8(%rcx), %rax
        leaq    1(%r8), %rdi
        movl    $1, %ecx
        xorl    %r14d, %r14d
        movq    %r11, -80(%rbp)
        movq    %r9, -152(%rbp)
        movq    -152(%rbp), %r9
        jmp     L466
        nopw    %cs:(%rax,%rax)
; │ @ transpose.jl within `transpose_f!`
L416:
        movq    -160(%rbp), %rcx
; │ @ transpose.jl:104 within `transpose_f!`
; │┌ @ range.jl:837 within `iterate`
        leaq    1(%rcx), %r8
; │└
        incq    %r14
        addq    $8, %rbx
        movq    -144(%rbp), %rdx
        addq    %rdx, %rsi
        addq    $8, %r11
        addq    %rdx, %rax
; │┌ @ range.jl:837 within `iterate`
; ││┌ @ promotion.jl:468 within `==`
        cmpq    %r9, %rcx
        movq    %r8, %rcx
        movq    %r12, %r8
; │└└
        je      L711
; │ @ transpose.jl within `transpose_f!`
L466:
        movq    %rcx, -160(%rbp)
        movl    $1, %r15d
        movq    %r8, %r12
; │ @ transpose.jl:104 within `transpose_f!`
        cmpq    $16, %r8
        jb      L672
; │ @ transpose.jl within `transpose_f!`
        movl    $1, %r15d
; │ @ transpose.jl:104 within `transpose_f!`
        cmpq    $1, -64(%rbp)
        jne     L672
; │ @ transpose.jl within `transpose_f!`
        movq    %r12, %r15
        imulq   %r14, %r15
        leaq    (%r12,%r15), %rcx
        movq    %r12, %r8
        movq    -88(%rbp), %rdx
        leaq    (%rdx,%rcx,8), %r12
        movq    -80(%rbp), %rcx
        leaq    (%rcx,%r14,8), %rcx
; │ @ transpose.jl:104 within `transpose_f!`
        cmpq    %r12, %rcx
        jae     L578
; │ @ transpose.jl within `transpose_f!`
        movq    -88(%rbp), %rcx
        leaq    (%rcx,%r15,8), %r12
        leaq    (%r8,%r14), %rcx
        movq    -80(%rbp), %rdx
        leaq    (%rdx,%rcx,8), %rcx
        movl    $1, %r15d
; │ @ transpose.jl:104 within `transpose_f!`
        cmpq    %rcx, %r12
        movq    %r8, %r12
        jb      L672
L578:
        leaq    1(%r10), %r15
        xorl    %r12d, %r12d
        nopl    (%rax)
; │┌ @ array.jl:862 within `getindex`
L592:
        vmovups -96(%rsi,%r12,8), %ymm0
        vmovups -64(%rsi,%r12,8), %ymm1
        vmovups -32(%rsi,%r12,8), %ymm2
        vmovups (%rsi,%r12,8), %ymm3
; │└
; │┌ @ array.jl:905 within `setindex!`
        vmovups %ymm0, -96(%rbx,%r12,8)
        vmovups %ymm1, -64(%rbx,%r12,8)
        vmovups %ymm2, -32(%rbx,%r12,8)
        vmovups %ymm3, (%rbx,%r12,8)
        addq    $16, %r12
        cmpq    %r12, %r10
        jne     L592
; │└
; │ @ array.jl within `transpose_f!`
        movq    %r8, %r12
; │ @ transpose.jl:104 within `transpose_f!`
        cmpq    %r10, %r8
        je      L416
        nopl    (%rax,%rax)
L672:
        leaq    -1(%r15), %rcx
        imulq   -64(%rbp), %rcx
        leaq    (%r11,%rcx,8), %rcx
        nopl    (%rax)
; │┌ @ array.jl:862 within `getindex`
L688:
        movq    (%rax,%r15,8), %rdx
; │└
; │┌ @ array.jl:905 within `setindex!`
        movq    %rdx, (%rcx)
; │└
; │┌ @ range.jl:837 within `iterate`
        incq    %r15
; ││┌ @ promotion.jl:468 within `==`
        addq    %r13, %rcx
        cmpq    %r15, %rdi
; │└└
        jne     L688
        jmp     L416
; │ @ transpose.jl within `transpose_f!`
L711:
        movq    -120(%rbp), %rax
        movq    -136(%rbp), %rcx
        movq    %rax, (%rcx)
        movq    -72(%rbp), %rax
; │ @ transpose.jl:111 within `transpose_f!`
        addq    $168, %rsp
        popq    %rbx
        popq    %rdi
        popq    %rsi
        popq    %r12
        popq    %r13
        popq    %r14
        popq    %r15
        popq    %rbp
        vzeroupper
        retq
; │ @ transpose.jl:97 within `transpose_f!`
; │┌ @ range.jl:1045 within `==` @ promotion.jl:468
L752:
        cmpq    %r8, %rbx
        je      L236
        jmp     L153
        nop
; └└
```

You can compare your work to whatever you want, but you do need to compare your work to existing work and explain why that does not meet your needs. Perhaps the main thing to emphasize here is that these other numerical package are not written in or otherwise directly usable from C++. You can then explicitly contrast with Eigen in that while it is written in C++, it's API differs from MATLAB. Software does not exist in vacuum. Others have sought to solve the problem of wanting MATLAB semantics in an open source and faster executing environment. Please acknowledge prior efforts and differentiate your effort from those efforts using the points you have outlined. I can only review that section after it has been written.
