### 2021-12-16 #3965 @mkitti
I just got my MSYS2 on Windows into decent shape, so I will try that again shortly.

We probably need more specific information regarding what compilers and what versions have been tested. It may also be necessary to specify to cmake the required the C++ standard to use.
