### 2021-12-16 #3965 @mkitti
See https://cmake.org/cmake/help/latest/prop_tgt/CXX_STANDARD.html
