### 2021-12-16 #3965 @mkitti
@jbytecode Could you outline the "State of the Field" requirements here? The only comparison is to the MATLAB, and I think there are other open source alternatives that try to address the stated problem, albeit in distinct ways. In reviewing the requirements with you, we determined that this perhaps does not need to be a distinct section, but the content does need to be clearly somewhere in the manuscript - likely part of "Statement of Need".
