### 2021-12-17 #3965 @jbytecode
State of the field can be a distinct section or the authors should describe how this software compares to other commonly-used packages in somewhere else in the paper, but yes, this comparison is somehow required.
