### 2021-12-17 #3965 @matthieuaussal
We just made this comparison between Matlab, Julia, Eigen and Castor and we planned to add this bench in the paper next tuesday, in a new "state of the field" part.

But you can already have a look at these attached codes.

Have a nice we

> Le 17 déc. 2021 à 16:46, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> State of the field can be a distinct section or the authors should describe how this software compares to other commonly-used packages in somewhere else in the paper, but yes, this comparison is somehow required.
>
