### 2021-12-21 4ea1ccf
Add comparison with other langages
