### 2021-12-21 #3965 @matthieuaussal
@mkitti:
First of all, we sincerely want to thank you for the quality of your review and the time you spend on it.

We added a point on the backslash operator after the example of the linear algebra part. The linsolve function that we propose is close to backslash, except that we do not deal specifically for triangular or symmetrical cases.

We agree with your remarks on setting up a conda package and clarifying the documentation. We will gradually update the site and propose a conda package to simplify the installation process.

We had already encountered this error message in demo_hmatrix.cpp, which comes from the default version of BLAS on Ubuntu. We fixed it by installing "open blas". We will specify it in the documentation.

We added in the summary and statement of field that Castor has been tested and validated on clang and gcc, in c ++ 14 standard.

We added State of field section, with comparison and discussion between Matlab, Julia, Eigen and Castor semantics.

We remain at your disposal for the rest,
Best regards,

Matthieu Aussal and Laurent Series
