### 2021-12-21 #3965 @mkitti
Thank you for clarifying the blas instructions. I will make another compilation attempt on Windows and Ubuntu later today.
