### 2021-12-21 #3965 @mkitti
Under MSys2, `cmake` has trouble finding `vtkIOOggTheora` while following the directions here:
https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/installation.html

```
$ cmake -G"MSYS Makefiles" -DBLAS_LIBRARIES="/mingw64/lib/libopenblas.a" -DLAPACK_LIBRARIES="/mingw64/lib/libopenblas.a" -DCBLAS_INCLUDE_DIR="/mingw64/include/OpenBLAS" ..
-- Building castor v1.0.0
-- BLAS and LAPACK found
-- BLAS_LIBRARIES : C:/msys64/mingw64/lib/libopenblas.a
-- LAPACK_LIBRARIES : C:/msys64/mingw64/lib/libopenblas.a
CMake Error at C:/msys64/mingw64/lib/cmake/vtk-8.2/vtkModuleAPI.cmake:140 (message):
  Requested modules not available:

    vtkIOOggTheora
Call Stack (most recent call first):
  C:/msys64/mingw64/lib/cmake/vtk-8.2/VTKConfig.cmake:143 (vtk_module_config)
  CMakeLists.txt:24 (find_package)

-- Configuring incomplete, errors occurred!
See also "C:/Users/kittisopikulm/source/repos/castor/build/CMakeFiles/CMakeOutput.log".
```

It appears this MSys2 commit may be relevant.
https://github.com/msys2/MINGW-packages/commit/08315cafe4dbd247b8ffb05182bc07cffe33f408#diff-dd2d765f56143d7820e2f73c88c2146aa0f0975fca62028d5421b2a49a5f2aaaL112

Notably, this line is removed:
```
    -DVTK_USE_SYSTEM_OGGTHEORA=ON \
```

These lines are added:
```
    -DVTK_USE_SYSTEM_OGG=ON \
    -DVTK_USE_SYSTEM_THEORA=ON \
```
