### 2021-12-22 #3965 @mkitti
If I disable VTK, I can get past the above errors. Then I run into errors such as these
```
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormbr.f.obj):sormbr.f:(.text+0x39a): undefined reference to `_gfortran_concat_string'
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormbr.f.obj):sormbr.f:(.text+0x5b1): undefined reference to `_gfortran_concat_string'
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormbr.f.obj):sormbr.f:(.text+0x759): undefined reference to `_gfortran_concat_string'
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormbr.f.obj):sormbr.f:(.text+0x7c1): undefined reference to `_gfortran_concat_string'
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormlq.f.obj):sormlq.f:(.text+0x2be): undefined reference to `_gfortran_concat_string'
C:/msys64/mingw64/bin/../lib/gcc/x86_64-w64-mingw32/11.2.0/../../../../x86_64-w64-mingw32/bin/ld.exe: C:/msys64/mingw64/lib/libopenblas.a(sormlq.f.obj):sormlq.f:(.text+0x941): more undefined references to `_gfortran_concat_string' follow
collect2.exe: error: ld returned 1 exit status
make[2]: *** [demo/CMakeFiles/demo_linalg.dir/build.make:157: demo/demo_linalg.exe] Error 1
make[1]: *** [CMakeFiles/Makefile2:352: demo/CMakeFiles/demo_linalg.dir/all] Error 2
make: *** [Makefile:136: all] Error 2
```

This suggests that `gfortran` is not getting linked correctly or at all. `demo/CMakeFiles/demo_linalg.dir/build.make:157` is as follows:

```
        cd /C/Users/kittisopikulm/source/repos/castor/build/demo && /C/msys64/mingw64/bin/g++.exe -Wl,--whole-archive CMakeFiles/demo_linalg.dir/objects.a -Wl,--no-whole-archive -o demo_linalg.exe -Wl,--out-implib,libdemo_linalg.dll.a -Wl,--major-image-version,0,--minor-image-version,0  /C/msys64/mingw64/lib/libopenblas.a /C/msys64/mingw64/lib/libvtkChartsCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkInfovisCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingContextOpenGL2-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingGL2PSOpenGL2-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingOpenGL2-8.2.dll.a /C/msys64/mingw64/lib/libvtkglew-8.2.dll.a /C/msys64/mingw64/lib/libvtkgl2ps-8.2.dll.a /C/msys64/mingw64/lib/libpng.dll.a /C/msys64/mingw64/lib/libvtkViewsContext2D-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingContext2D-8.2.dll.a /C/msys64/mingw64/lib/libvtkViewsCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkInteractionWidgets-8.2.dll.a /C/msys64/mingw64/lib/libvtkInteractionStyle-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersExtraction-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersStatistics-8.2.dll.a /C/msys64/mingw64/lib/libvtkImagingFourier-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersHybrid-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersModeling-8.2.dll.a /C/msys64/mingw64/lib/libvtkImagingGeneral-8.2.dll.a /C/msys64/mingw64/lib/libvtkImagingSources-8.2.dll.a /C/msys64/mingw64/lib/libvtkImagingHybrid-8.2.dll.a /C/msys64/mingw64/lib/libvtkIOImage-8.2.dll.a /C/msys64/mingw64/lib/libvtkDICOMParser-8.2.dll.a /C/msys64/mingw64/lib/libvtkmetaio-8.2.dll.a /C/msys64/mingw64/lib/libz.dll.a -lcomctl32 -lwsock32 /C/msys64/mingw64/lib/libjpeg.dll.a /C/msys64/mingw64/lib/libtiff.dll.a /C/msys64/mingw64/lib/libvtkRenderingAnnotation-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingFreeType-8.2.dll.a /C/msys64/mingw64/lib/libfreetype.dll.a /C/msys64/mingw64/lib/libvtkImagingColor-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingVolume-8.2.dll.a /C/msys64/mingw64/lib/libvtkImagingCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkRenderingCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonColor-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersGeometry-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersSources-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersGeneral-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonComputationalGeometry-8.2.dll.a /C/msys64/mingw64/lib/libvtkFiltersCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkIOXML-8.2.dll.a /C/msys64/mingw64/lib/libdouble-conversion.dll.a /C/msys64/mingw64/lib/liblz4.dll.a /C/msys64/mingw64/lib/libvtkIOXMLParser-8.2.dll.a /C/msys64/mingw64/lib/libexpat.dll.a /C/msys64/mingw64/lib/libvtkIOPLY-8.2.dll.a /C/msys64/mingw64/lib/libvtkIOLegacy-8.2.dll.a /C/msys64/mingw64/lib/libvtkIOCore-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonExecutionModel-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonDataModel-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonTransforms-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonMisc-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonMath-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonSystem-8.2.dll.a /C/msys64/mingw64/lib/libvtkCommonCore-8.2.dll.a /C/msys64/mingw64/lib/libtbb.dll.a /C/msys64/mingw64/lib/libvtksys-8.2.dll.a -lws2_32 -lPsapi /C/msys64/mingw64/lib/libvtklzma-8.2.dll.a -lgdi32 -lkernel32 -luser32 -lgdi32 -lwinspool -lshell32 -lole32 -loleaut32 -luuid -lcomdlg32 -ladvapi32
```

If I manually append `-lgfortran` to the above line, then it compiles fine. I have not figured out how to convince CMake to do this for me. Adding `gfortran` to `target_link_libraries` for the targets adds it too early, which does not work since it needs to be added statically.

The targets affected are
* demo/demo_hmatrix.exe
* demo/demo_linalg.exe
* misc/perf_hmat.exe
