### 2021-12-22 #3965 @mkitti
When running `perf_hmat.exe`, I got a segmentation fault the first time I ran it:

```
$ ./perf_hmat.exe
+=====================+
|      INITIALIZE     |
+=====================+
============================ DOCUMENTATION ============================
Source file '/usr/local/include/castor/matrix.hpp' not found.
Source file '/usr/local/include/castor/hmatrix.hpp' not found.
Function "hmatrix" not found in file(s):
  /usr/local/include/castor/matrix.hpp
  /usr/local/include/castor/hmatrix.hpp
=======================================================================
---> DISCRETIZATION X <---
Bounding box data :
  Minimum :    -0.99763     -0.99933     -1.00000
  Center  :  5.4087e-04   2.3382e-04            0
  Maximum :     0.99871      0.99980      1.00000
  Edge    :     1.99634      1.99914      2.00000
---> DISCRETIZATION Y <---
Bounding box data :
  Minimum :    -0.99763     -0.99933     -1.00000
  Center  :  5.4087e-04   2.3382e-04            0
  Maximum :     0.99871      0.99980      1.00000
  Edge    :     1.99634      1.99914      2.00000
0
---> FULL <---
Elapsed time is 0.130748 seconds.
---> FULL TO H-MATRIX <---
Elapsed time is 0.0560277 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (11.3149 MB)
  Leaf : 85h 110c 146f
  Rate : 70.7179 %
  tol  : 0.001
0.000575456
---> SPARSE TO H-MATRIX <---
Elapsed time is 3.03438 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (16 MB)
  Leaf : 85h 0c 256f
  Rate : 100 %
  tol  : 0.001
0
---> FULL MATRIX REFERENCE <---
Elapsed time is 0.0122054 seconds.
---> H-MATRIX <---
Elapsed time is 0.126789 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (11.3149 MB)
  Leaf : 85h 110c 146f
  Rate : 70.7179 %
  tol  : 0.001
0.000202287
---> FACTO LU <---
Elapsed time is 1.51904 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (6.272 MB)
  Leaf : 50h 70c 81f
  Rate : 39.2 %
  tol  : 0.001
H-matrix 1000x1000 of type 'St7complexIdE' (6.26603 MB)
  Leaf : 50h 70c 81f
  Rate : 39.1627 %
  tol  : 0.001
0.000209073
---> GMRES + PRECOND LU <---
Segmentation fault
```

On a subsequent execution it worked.

```
$ ./perf_hmat.exe
+=====================+
|      INITIALIZE     |
+=====================+
============================ DOCUMENTATION ============================
Source file '/usr/local/include/castor/matrix.hpp' not found.
Source file '/usr/local/include/castor/hmatrix.hpp' not found.
Function "hmatrix" not found in file(s):
  /usr/local/include/castor/matrix.hpp
  /usr/local/include/castor/hmatrix.hpp
=======================================================================
---> DISCRETIZATION X <---
Bounding box data :
  Minimum :    -0.99763     -0.99933     -1.00000
  Center  :  5.4087e-04   2.3382e-04            0
  Maximum :     0.99871      0.99980      1.00000
  Edge    :     1.99634      1.99914      2.00000
---> DISCRETIZATION Y <---
Bounding box data :
  Minimum :    -0.99763     -0.99933     -1.00000
  Center  :  5.4087e-04   2.3382e-04            0
  Maximum :     0.99871      0.99980      1.00000
  Edge    :     1.99634      1.99914      2.00000
0
---> FULL <---
Elapsed time is 0.12465 seconds.
---> FULL TO H-MATRIX <---
Elapsed time is 0.0565889 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (11.3149 MB)
  Leaf : 85h 110c 146f
  Rate : 70.7179 %
  tol  : 0.001
0.000575456
---> SPARSE TO H-MATRIX <---
Elapsed time is 3.03659 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (16 MB)
  Leaf : 85h 0c 256f
  Rate : 100 %
  tol  : 0.001
0
---> FULL MATRIX REFERENCE <---
Elapsed time is 0.0133527 seconds.
---> H-MATRIX <---
Elapsed time is 0.126242 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (11.3149 MB)
  Leaf : 85h 110c 146f
  Rate : 70.7179 %
  tol  : 0.001
0.000202287
---> FACTO LU <---
Elapsed time is 1.5923 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (6.272 MB)
  Leaf : 50h 70c 81f
  Rate : 39.2 %
  tol  : 0.001
H-matrix 1000x1000 of type 'St7complexIdE' (6.26603 MB)
  Leaf : 50h 70c 81f
  Rate : 39.1627 %
  tol  : 0.001
0.000209073
---> GMRES + PRECOND LU <---
Elapsed time is 0.93759 seconds.
H-matrix 1000x1000 of type 'St7complexIdE' (5.48235 MB)
  Leaf : 50h 70c 81f
  Rate : 34.2647 %
  tol  : 0.1
H-matrix 1000x1000 of type 'St7complexIdE' (5.4864 MB)
  Leaf : 50h 70c 81f
  Rate : 34.29 %
  tol  : 0.1
0.0200213
Start GMRES using MGCR implementation (Multiple Generalized Conjugate Residual).
GMRES converged at iteration 8 to a solution with relative residual (0.000199289,0)
and relative error 0.000387232.
done !
```
