### 2021-12-22 #3965 @mkitti
If I alter `castor/CMakeLists.txt` by commenting out the `vtkIOOggTheora` line,
```
$ git diff
diff --git a/CMakeLists.txt b/CMakeLists.txt
index b6fa1bf..4c49151 100644
--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -39,7 +39,7 @@ find_package(VTK COMPONENTS
   vtkIOImage
   vtkIOPLY
   vtkIOLegacy
-  vtkIOOggTheora
+  #vtkIOOggTheora
   vtkFiltersGeometry
   QUIET
 )
```

then I get the following output:

```
$ cmake -G"MSYS Makefiles" -DBLAS_LIBRARIES="/mingw64/lib/libopenblas.a" -DLAPACK_LIBRARIES="/mingw64/lib/libopenblas.a" -DCBLAS_INCLUDE_DIR="/mingw64/include/OpenBLAS" ..
-- Building castor v1.0.0
-- BLAS and LAPACK found
-- BLAS_LIBRARIES : C:/msys64/mingw64/lib/libopenblas.a
-- LAPACK_LIBRARIES : C:/msys64/mingw64/lib/libopenblas.a
-- Find VTK version : 8.2.0
-- VTK_LIBRARIES    : vtkChartsCore;vtkCommonColor;vtkCommonCore;vtksys;vtkCommonDataModel;vtkCommonMath;vtkCommonMisc;vtkCommonSystem;vtkCommonTransforms;vtkCommonExecutionModel;vtkFiltersGeneral;vtkCommonComputationalGeometry;vtkFiltersCore;vtkInfovisCore;vtkFiltersExtraction;vtkFiltersStatistics;vtkImagingFourier;vtkImagingCore;vtkRenderingContext2D;vtkRenderingCore;vtkFiltersGeometry;vtkFiltersSources;vtkRenderingFreeType;C:/msys64/mingw64/lib/libfreetype.dll.a;C:/msys64/mingw64/lib/libz.dll.a;vtkInteractionStyle;vtkRenderingContextOpenGL2;vtkRenderingOpenGL2;vtkglew;vtkRenderingGL2PSOpenGL2;vtkgl2ps;C:/msys64/mingw64/lib/libpng.dll.a;vtkViewsContext2D;vtkViewsCore;vtkInteractionWidgets;vtkFiltersHybrid;vtkImagingSources;vtkFiltersModeling;vtkImagingColor;vtkImagingGeneral;vtkImagingHybrid;vtkIOImage;vtkDICOMParser;vtkmetaio;C:/msys64/mingw64/lib/libjpeg.dll.a;C:/msys64/mingw64/lib/libtiff.dll.a;vtkRenderingAnnotation;vtkRenderingVolume;vtkIOXML;vtkIOCore;C:/msys64/mingw64/lib/libdouble-conversion.dll.a;C:/msys64/mingw64/lib/liblz4.dll.a;vtklzma;vtkIOXMLParser;C:/msys64/mingw64/lib/libexpat.dll.a;vtkIOPLY;vtkIOLegacy
CMake Error at demo/CMakeLists.txt:30 (vtk_module_autoinit):
  Unknown CMake command "vtk_module_autoinit".

-- Configuring incomplete, errors occurred!
See also "C:/source/repos/castor/build/CMakeFiles/CMakeOutput.log".
```

I am not sure if `vtk_module_autoinit` is available with VTK 8.2.
