### 2021-12-22 #3965 @mkitti
I note that examples for MATLAB and Julia do not have a "warmup" step for the timings, which is necessary for a fair comparison because both languages use a form of just-in-time compilation. Julia is using [ORCv2 JIT from LLVM](https://llvm.org/docs/ORCv2.html). Perhaps the alternative would be adding a timing step to measure compilation time with Castor?

In the Julia case, this is quite obvious since `@time` reports the percent of compilation time.
```julia
julia> using LinearAlgebra

julia> @time test()
4×3 Matrix{Int64}:
  1   2   3
  4   5   6
  7   8   9
 10  11  12
4×3 Matrix{Int64}:
 -1  0  0
  0  4  0
  0  0  8
  0  0  0
4×1 Matrix{Int64}:
 -1
  4
  8
  0
4×3 Matrix{Int64}:
 1  0  0
 0  4  0
 0  0  8
 0  0  0
4×3 Matrix{Int64}:
 -1  0  0
  0  0  0
  0  0  0
  0  4  8
4×4 Matrix{Int64}:
 1   0   0  0
 0  16   0  0
 0   0  64  0
 0   0   0  0
  0.127030 seconds (89.62 k allocations: 5.542 MiB, 90.51% compilation time)

```
Subsequent calls to `test()` are much faster since they do not include compilation time.

```
julia> @time test();
...
  0.006393 seconds (4.13 k allocations: 303.078 KiB)
```

Similarly, the MATLAB code could be refactored into a function, similar to the Julia code. Subsequent evaluations of the MATLAB function are also faster.

```matlab
tic; test(); toc
disp("done.")

function test()

M = [1 2 3; 4 5 6; 7 8 9; ...
    10 11 13];
disp(M)
M = (M - 1) .* eye(size(M));

M(1,1) = -1;
M([2,3], 1) = -1;
M(4,:) = -1;
disp(M);

disp(sum(M,2));
disp(abs(M));
disp(sort(M,1));
disp(M*M');

end
```

The output is as follows.
```matlab
>> demoscript
     1     2     3
     4     5     6
     7     8     9
    10    11    12

    -1     0     0
    -1     4     0
    -1     0     8
    -1    -1    -1

    -1
     3
     7
    -3

     1     0     0
     1     4     0
     1     0     8
     1     1     1

    -1    -1    -1
    -1     0     0
    -1     0     0
    -1     4     8

     1     1     1     1
     1    17     1    -3
     1     1    65    -7
     1    -3    -7     3

Elapsed time is 0.001682 seconds.
done.

...

Elapsed time is 0.000731 seconds.
done.
```

For benchmarking purposes, the convention is to use [`timeit`](https://www.mathworks.com/help/matlab/ref/timeit.html) or [`@btime`](https://juliaci.github.io/BenchmarkTools.jl/stable/manual/#Benchmarking-basics) in BenchmarkingTools.jl in MATLAB and Julia respectively. Furthermore, Julia precompilation is available by using a module or full ahead-of-time compilation is available via [PackageCompiler.jl](https://github.com/JuliaLang/PackageCompiler.jl).

Since the authors do not actually reporting benchmark times in the article, perhaps the purpose of the timings is not benchmarking. In that case, it may be useful to note the purpose of timing code execution. As stated above, an alternative might be to include a mechanism to time of the compilation process for Castor. Perhaps another option is just to remove the timings since they are not directly comparable nor do the authors report those timing results.
