### 2021-12-22 #3965 @mkitti
Another point is that the authors discuss MATLAB and Julia as interpreted languages. However, both languages do employ some form of just-in-time compilation. [The MATLAB Execution Engine has been part of MATLAB since R2015B](https://www.mathworks.com/products/matlab/matlab-execution-engine.html). Julia uses LLVM ORCv2 JIT as mentioned above. The machine code for Julia can be printed as follows.

```julia
julia> @code_native test()
        .text
; ┌ @ REPL[2]:1 within `test'
        pushq   %rbp
        movq    %rsp, %rbp
        pushq   %r15
        pushq   %r14
        pushq   %r13
        pushq   %r12
        pushq   %rbx
        andq    $-32, %rsp
        subq    $480, %rsp                      # imm = 0x1E0
        movq    %rsp, %rbx
        vxorps  %xmm0, %xmm0, %xmm0
        vmovaps %ymm0, 128(%rbx)
...
```

<summary>Full x86_64 assembly of `test()` from Julia</summary>

julia> @code_native test()
        .text
; ┌ @ REPL[2]:1 within `test'
        pushq   %rbp
        movq    %rsp, %rbp
        pushq   %r15
        pushq   %r14
        pushq   %r13
        pushq   %r12
        pushq   %rbx
        andq    $-32, %rsp
        subq    $480, %rsp                      # imm = 0x1E0
        movq    %rsp, %rbx
        vxorps  %xmm0, %xmm0, %xmm0
        vmovaps %ymm0, 128(%rbx)
        movq    $0, 160(%rbx)
        movq    %fs:0, %r12
; │ @ REPL[2]:3 within `test'
        movq    $12, 128(%rbx)
        movq    -32768(%r12), %rax
        movq    %rax, 136(%rbx)
        leaq    128(%rbx), %rax
        movq    %rax, -32768(%r12)
        movabsq $.rodata.cst32, %rax
        vmovaps (%rax), %ymm0
        vmovups %ymm0, 256(%rbx)
        movabsq $139848898793920, %rax          # imm = 0x7F311BEF51C0
        vmovaps (%rax), %ymm0
        vmovups %ymm0, 288(%rbx)
        movabsq $139848898793952, %rax          # imm = 0x7F311BEF51E0
        vmovdqa (%rax), %ymm0
        vmovdqu %ymm0, 320(%rbx)
        movabsq $139850219012896, %rax          # imm = 0x7F316AA04320
        movq    %rax, 352(%rbx)
        movabsq $jl_invoke, %rax
        movabsq $jl_system_image_data, %rdi
        leaq    256(%rbx), %r14
        movabsq $139850244403648, %rcx          # imm = 0x7F316C23B1C0
        movq    %r14, %rsi
        movl    $13, %edx
        vzeroupper
        callq   *%rax
        movq    %rax, %r15
        movq    %rax, 144(%rbx)
; │ @ REPL[2]:5 within `test'
        movq    %rax, 256(%rbx)
        movabsq $display, %rax
        movabsq $jl_system_image_data, %rdi
        movq    %r14, %rsi
        movl    $1, %edx
        callq   *%rax
; │ @ REPL[2]:6 within `test'
; │┌ @ array.jl:134 within `size'
        movq    24(%r15), %r14
        movq    %r15, %r13
        movq    32(%r15), %r15
        movq    %r14, 440(%rbx)
        movq    %r15, 448(%rbx)
; │└
; │┌ @ uniformscaling.jl:473 within `Array'
        movabsq $.rodata, %rdi
        movabsq $Array, %rax
        leaq    440(%rbx), %rsi
        callq   *%rax
; │└
; │┌ @ broadcast.jl:883 within `materialize'
; ││┌ @ broadcast.jl:266 within `instantiate'
; │││┌ @ broadcast.jl:484 within `combine_axes'
; ││││┌ @ broadcast.jl:207 within `axes'
; │││││┌ @ broadcast.jl:209 within `_axes'
; ││││││┌ @ broadcast.jl:484 within `combine_axes'
; │││││││┌ @ broadcast.jl:489 within `broadcast_shape'
; ││││││││┌ @ broadcast.jl:493 within `_bcs'
        movq    %r14, 424(%rbx)
        movq    %r15, 432(%rbx)
; ││││└└└└└
; ││││┌ @ abstractarray.jl:89 within `axes'
; │││││┌ @ array.jl:134 within `size'
        movq    24(%rax), %rcx
        movq    32(%rax), %rdx
        movq    %rcx, 208(%rbx)
; │││││└
; │││││┌ @ tuple.jl:214 within `map'
        movq    %rcx, 408(%rbx)
        movq    %rdx, 200(%rbx)
        movq    %rdx, 416(%rbx)
        movq    %rax, 88(%rbx)
        movq    %rax, 160(%rbx)
; ││││└└
; ││││┌ @ broadcast.jl:489 within `broadcast_shape'
        movabsq $_bcs, %rax
        leaq    240(%rbx), %rdi
        leaq    424(%rbx), %rsi
        leaq    408(%rbx), %rdx
        callq   *%rax
; ││└└└
; ││┌ @ broadcast.jl:908 within `copy'
; │││┌ @ broadcast.jl:196 within `similar' @ broadcast.jl:197 @ abstractarray.jl:784 @ abstractarray.jl:785
; ││││┌ @ boot.jl:465 within `Array' @ boot.jl:458 @ boot.jl:450
        movq    240(%rbx), %r14
        movq    248(%rbx), %rdx
        movabsq $jl_system_image_data, %rdi
        movq    %r14, %rsi
        movq    %rdx, 48(%rbx)
        movabsq $139850550859553, %rax          # imm = 0x7F317E67D721
        callq   *%rax
        movq    %rax, %rcx
; │││└└
; │││┌ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:972
; ││││┌ @ abstractarray.jl:89 within `axes'
; │││││┌ @ array.jl:134 within `size'
        movq    24(%rax), %rax
        movq    32(%rcx), %rdx
; ││││└└
; ││││┌ @ tuple.jl:362 within `=='
; │││││┌ @ tuple.jl:366 within `_eq'
; ││││││┌ @ range.jl:802 within `=='
; │││││││┌ @ range.jl:519 within `isempty'
; ││││││││┌ @ operators.jl:305 within `>'
; │││││││││┌ @ int.jl:83 within `<'
        testq   %rax, %rax
; │││││││└└└
        je      L504
; ││││││└
; ││││││┌ @ range.jl:803 within `=='
; │││││││┌ @ range.jl:816 within `_has_length_one'
; ││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, %rax
; │││││││└└
        jne     L514
; │││││││┌ @ range.jl:816 within `_has_length_one'
; ││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, %r14
; │││││└└└└
; │││││┌ @ tuple.jl:367 within `_eq'
        je      L523
        jmp     L3209
; │││││└
; │││││┌ @ tuple.jl:366 within `_eq'
; ││││││┌ @ range.jl:802 within `=='
; │││││││┌ @ range.jl:519 within `isempty'
; ││││││││┌ @ operators.jl:305 within `>'
; │││││││││┌ @ int.jl:83 within `<'
L504:
        testq   %r14, %r14
; │││││└└└└└
; │││││┌ @ tuple.jl:367 within `_eq'
        jle     L523
        jmp     L3209
; │││││└
; │││││┌ @ tuple.jl:366 within `_eq'
; ││││││┌ @ range.jl:804 within `==' @ promotion.jl:410
L514:
        cmpq    %r14, %rax
; │││││└└
; │││││┌ @ tuple.jl:367 within `_eq'
        jne     L3209
; │││││└
; │││││┌ @ tuple.jl:372 within `_eq' @ tuple.jl:366
; ││││││┌ @ range.jl:802 within `=='
; │││││││┌ @ range.jl:519 within `isempty'
; ││││││││┌ @ operators.jl:305 within `>'
; │││││││││┌ @ int.jl:83 within `<'
L523:
        testq   %rdx, %rdx
; │││││││└└└
        je      L550
; ││││││└
; ││││││┌ @ range.jl:803 within `=='
; │││││││┌ @ range.jl:816 within `_has_length_one'
; ││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, %rdx
; │││││││└└
        jne     L3199
; │││││││┌ @ range.jl:816 within `_has_length_one'
; ││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, 48(%rbx)
; ││││└└└└└
        je      L561
        jmp     L3209
; ││││┌ @ tuple.jl:362 within `=='
; │││││┌ @ tuple.jl:372 within `_eq' @ tuple.jl:366
; ││││││┌ @ range.jl:802 within `=='
; │││││││┌ @ range.jl:519 within `isempty'
; ││││││││┌ @ operators.jl:305 within `>'
; │││││││││┌ @ int.jl:83 within `<'
L550:
        cmpq    $0, 48(%rbx)
; ││││└└└└└└
        jg      L3209
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:0
L561:
        movq    %rdx, 120(%rbx)
        movq    %rax, 96(%rbx)
        movabsq $139850550859553, %rax          # imm = 0x7F317E67D721
        movb    -232606248(%rax), %r15b
        movl    %r15d, %eax
        andb    $1, %al
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:980
; ││││┌ @ broadcast.jl:963 within `preprocess'
; │││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││┌ @ broadcast.jl:963 within `preprocess'
; │││││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││││┌ @ broadcast.jl:957 within `broadcast_unalias'
        cmpq    %r13, %rcx
        movq    %rcx, 72(%rbx)
        movb    %al, 31(%rbx)
        je      L651
        testb   %al, %al
        jne     L651
; ││││││││││┌ @ abstractarray.jl:1349 within `unalias'
; │││││││││││┌ @ abstractarray.jl:1384 within `mightalias'
; ││││││││││││┌ @ abstractarray.jl:1408 within `dataids'
; │││││││││││││┌ @ abstractarray.jl:1116 within `pointer'
; ││││││││││││││┌ @ pointer.jl:65 within `unsafe_convert'
        movq    (%rcx), %rax
; ││││││││││││└└└
; ││││││││││││┌ @ abstractarray.jl:1391 within `_isdisjoint'
; │││││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    (%r13), %rax
; │││││││││││└└└└
        jne     L651
; ││││││││││└
; ││││││││││┌ @ abstractarray.jl within `unalias'
        movq    %rcx, 152(%rbx)
; ││││││││││└
; ││││││││││┌ @ abstractarray.jl:1349 within `unalias'
; │││││││││││┌ @ abstractarray.jl:1365 within `unaliascopy'
; ││││││││││││┌ @ array.jl:349 within `copy'
        movabsq $139850550859553, %rax          # imm = 0x7F317E67D721
        addq    $204, %rax
        movq    %r13, %rdi
        callq   *%rax
        movq    72(%rbx), %rcx
        movq    %rax, %r13
; │││││││││└└└└
; │││││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││││┌ @ abstractarray.jl:89 within `axes'
; ││││││││││││┌ @ array.jl:134 within `size'
L651:
        movq    24(%r13), %rax
        movq    %rax, 112(%rbx)
        movq    32(%r13), %rax
        movq    %rax, 224(%rbx)
; ││││││└└└└└└└
; ││││││ @ broadcast.jl:966 within `preprocess_args' @ broadcast.jl:967
; ││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││┌ @ broadcast.jl:957 within `broadcast_unalias'
; ││││││││┌ @ abstractarray.jl:1349 within `unalias'
; │││││││││┌ @ abstractarray.jl:1384 within `mightalias'
        testb   $1, %r15b
        movq    %r12, 216(%rbx)
        movq    %r13, 232(%rbx)
        jne     L785
        movabsq $139850550859553, %rax          # imm = 0x7F317E67D721
        testb   $1, -222780056(%rax)
        movq    96(%rbx), %r15
        jne     L789
; ││││││││││┌ @ abstractarray.jl:1408 within `dataids'
; │││││││││││┌ @ abstractarray.jl:1116 within `pointer'
; ││││││││││││┌ @ pointer.jl:65 within `unsafe_convert'
        movq    (%rcx), %rax
        movq    88(%rbx), %rdi
; ││││││││││└└└
; ││││││││││┌ @ abstractarray.jl:1391 within `_isdisjoint'
; │││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    (%rdi), %rax
; │││││││││└└└└
        jne     L789
; ││││││││└
; ││││││││┌ @ abstractarray.jl within `unalias'
        movq    %rcx, 152(%rbx)
        movq    %r13, 144(%rbx)
; ││││││││└
; ││││││││┌ @ abstractarray.jl:1349 within `unalias'
; │││││││││┌ @ abstractarray.jl:1365 within `unaliascopy'
; ││││││││││┌ @ array.jl:349 within `copy'
        movabsq $139850550859553, %rax          # imm = 0x7F317E67D721
        addq    $204, %rax
        callq   *%rax
; │││││││└└└└
; │││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││┌ @ abstractarray.jl:89 within `axes'
; ││││││││││┌ @ array.jl:134 within `size'
        movq    24(%rax), %rcx
        movq    %rcx, 208(%rbx)
        movq    %rax, 88(%rbx)
        movq    32(%rax), %rax
        movq    %rax, 200(%rbx)
        jmp     L789
; ││││└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ array.jl:0
L785:
        movq    96(%rbx), %r15
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:983
; ││││┌ @ simdloop.jl:70 within `macro expansion'
; │││││┌ @ multidimensional.jl:374 within `iterate'
; ││││││┌ @ tuple.jl:232 within `map'
; │││││││┌ @ range.jl:1090 within `in'
; ││││││││┌ @ int.jl:442 within `<='
L789:
        cmpq    $0, 48(%rbx)
; │││││└└└└
        jle     L2739
; ││││└
; ││││┌ @ simdloop.jl:72 within `macro expansion'
; │││││┌ @ int.jl within `<'
        testq   %r14, %r14
; │││││└
        jle     L2739
; ││││└
; ││││┌ @ simdloop.jl within `macro expansion'
        movq    112(%rbx), %rax
; ││││└
; ││││┌ @ simdloop.jl:75 within `macro expansion'
        leaq    (,%rax,8), %rax
        movq    %rax, 192(%rbx)
        movq    %r14, %r13
        andq    $-16, %r13
        movl    $1, %r11d
        movl    $12, %eax
        movq    %rax, 64(%rbx)
        vpcmpeqd        %xmm8, %xmm8, %xmm8
        movabsq $.rodata.cst4, %rax
        vpbroadcastd    (%rax), %xmm9
        movabsq $.rodata.cst16, %rax
        vmovdqa (%rax), %xmm11
        vpcmpeqd        %ymm10, %ymm10, %ymm10
        xorl    %eax, %eax
        movq    %rax, 80(%rbx)
        xorl    %r10d, %r10d
        movq    %r13, 104(%rbx)
        nopw    %cs:(%rax,%rax)
; │││││┌ @ int.jl within `<'
L912:
        testq   %r14, %r14
; │││││└
        jle     L1392
; ││││└
; ││││┌ @ simdloop.jl within `macro expansion'
        movq    96(%rbx), %rdi
        movq    %r10, %r13
        imulq   %r10, %rdi
        leaq    (%r14,%rdi), %rax
        movq    %rax, 56(%rbx)
        leaq    -1(%r11), %r8
; ││││└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:980
; ││││┌ @ broadcast.jl:963 within `preprocess'
; │││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││┌ @ broadcast.jl:963 within `preprocess'
; │││││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││││┌ @ broadcast.jl:565 within `shapeindexer'
; ││││││││││││┌ @ broadcast.jl:569 within `_newindexer' @ broadcast.jl:570
; │││││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, 224(%rbx)
; ││││└└└└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:983
; ││││┌ @ simdloop.jl:77 within `macro expansion' @ broadcast.jl:984
; │││││┌ @ broadcast.jl:575 within `getindex'
; ││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││││┌ @ broadcast.jl:614 within `_broadcast_getindex'
; │││││││││││┌ @ multidimensional.jl:637 within `getindex' @ array.jl:0
        movq    %r8, %rdx
        movl    $0, %eax
        cmoveq  %rax, %rdx
        movq    112(%rbx), %rsi
        movq    %rsi, %r12
        imulq   %rdx, %r12
; ││││└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:980
; ││││┌ @ broadcast.jl:963 within `preprocess'
; │││││┌ @ broadcast.jl:966 within `preprocess_args' @ broadcast.jl:967
; ││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││┌ @ broadcast.jl:565 within `shapeindexer'
; ││││││││││┌ @ broadcast.jl:569 within `_newindexer' @ broadcast.jl:570
; │││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, 200(%rbx)
; ││││└└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:983
; ││││┌ @ simdloop.jl:77 within `macro expansion' @ broadcast.jl:984
; │││││┌ @ broadcast.jl:575 within `getindex'
; ││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:644 within `_getindex' @ broadcast.jl:645
; ││││││││┌ @ broadcast.jl:614 within `_broadcast_getindex'
; │││││││││┌ @ multidimensional.jl:637 within `getindex' @ array.jl:0
        cmoveq  %rax, %r8
        movq    208(%rbx), %rcx
        imulq   %rcx, %r8
        movq    232(%rbx), %rax
        movq    (%rax), %r10
        movq    88(%rbx), %rax
        movq    (%rax), %r15
        movq    72(%rbx), %rax
        movq    (%rax), %r9
; ││││└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:980
; ││││┌ @ broadcast.jl:963 within `preprocess'
; │││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││┌ @ broadcast.jl:963 within `preprocess'
; │││││││┌ @ broadcast.jl:966 within `preprocess_args'
; ││││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││││┌ @ broadcast.jl:565 within `shapeindexer'
; ││││││││││││┌ @ broadcast.jl:570 within `_newindexer'
; │││││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, %rsi
        movq    %r9, 40(%rbx)
; ││││└└└└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:983
; ││││┌ @ simdloop.jl:75 within `macro expansion'
        jne     L1136
; ││││└
; ││││┌ @ simdloop.jl:77 within `macro expansion' @ broadcast.jl:984
; │││││┌ @ broadcast.jl:575 within `getindex'
; ││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││││┌ @ broadcast.jl:614 within `_broadcast_getindex'
; │││││││││││┌ @ multidimensional.jl:637 within `getindex' @ array.jl:0
        leaq    (%r10,%r12,8), %r12
; ││││└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:980
; ││││┌ @ broadcast.jl:963 within `preprocess'
; │││││┌ @ broadcast.jl:966 within `preprocess_args' @ broadcast.jl:967
; ││││││┌ @ broadcast.jl:964 within `preprocess'
; │││││││┌ @ broadcast.jl:615 within `extrude'
; ││││││││┌ @ broadcast.jl:564 within `newindexer'
; │││││││││┌ @ broadcast.jl:565 within `shapeindexer'
; ││││││││││┌ @ broadcast.jl:570 within `_newindexer'
; │││││││││││┌ @ operators.jl:204 within `!='
; ││││││││││││┌ @ promotion.jl:410 within `=='
        cmpq    $1, %rcx
        movq    %r10, %rcx
; ││││└└└└└└└└└
; ││││ @ broadcast.jl:936 within `copyto!' @ broadcast.jl:983
; ││││┌ @ simdloop.jl:75 within `macro expansion'
        jne     L1235
; ││││└
; ││││┌ @ simdloop.jl:77 within `macro expansion' @ broadcast.jl:984
; │││││┌ @ broadcast.jl:575 within `getindex'
; ││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:644 within `_getindex' @ broadcast.jl:645
; ││││││││┌ @ broadcast.jl:614 within `_broadcast_getindex'
; │││││││││┌ @ multidimensional.jl:637 within `getindex' @ array.jl:0
        leaq    (%r15,%r8), %rax
; ││││└└└└└└
; ││││┌ @ simdloop.jl:75 within `macro expansion'
        cmpq    $16, %r14
        movq    %r13, %r10
        jae     L1444
; ││││└
; ││││┌ @ simdloop.jl within `macro expansion'
        xorl    %ecx, %ecx
L1068:
        movq    80(%rbx), %rdx
        movq    40(%rbx), %rsi
; ││││└
; ││││┌ @ simdloop.jl:75 within `macro expansion'
        leaq    (%rsi,%rdx,8), %rdx
        nopl    (%rax,%rax)
; ││││└
; ││││┌ @ simdloop.jl:77 within `macro expansion' @ broadcast.jl:984
; │││││┌ @ broadcast.jl:575 within `getindex'
; ││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││┌ @ broadcast.jl:620 within `_broadcast_getindex'
; │││││││││┌ @ broadcast.jl:644 within `_getindex'
; ││││││││││┌ @ broadcast.jl:614 within `_broadcast_getindex'
; │││││││││││┌ @ multidimensional.jl:637 within `getindex' @ array.jl:806
L1088:
        movq    (%r12), %rsi
; ││││││││└└└└
; ││││││││┌ @ broadcast.jl:621 within `_broadcast_getindex'
; │││││││││┌ @ broadcast.jl:648 within `_broadcast_getindex_evalf'
; ││││││││││┌ @ int.jl:86 within `-'
        decq    %rsi
; ││││││└└└└└
; ││││││┌ @ broadcast.jl:621 within `_broadcast_getindex'
; │││││││┌ @ broadcast.jl:648 within `_broadcast_getindex_evalf'
; ││││││││┌ @ int.jl:921 within `*'
; │││││││││┌ @ int.jl:474 within `rem'
; ││││││││││┌ @ number.jl:7 within `convert'
; │││││││││││┌ @ boot.jl:752 within `Int64'
; ││││││││││││┌ @ boot.jl:679 within `toInt64'
        movzbl  (%rax), %edi
        andl    $1, %edi
; ││││││││└└└└└
; ││││││││┌ @ int.jl:923 within `*' @ int.jl:88
        imulq   %rsi, %rdi
; │││││└└└└
; │││││┌ @ multidimensional.jl:639 within `setindex!' @ array.jl:845
        movq    %rdi, (%rdx,%rcx,8)
; ││││└└
; ││││┌ @ simdloop.jl:78 within `macro expansion'
; │││││┌ @ int.jl:87 within `+'
        incq    %rcx
; ││││└└
; ││││┌ @ simdloop.jl:75 within `macro expansion'
; │││││┌ @ int.jl:83 within `<'
        cmpq    %rcx, %r14
; │││││└
        jne     L1088
        jmp     L1392
        nopw    %cs:(%rax,%rax)
; ││││└
; ││││ @ broadcast.jl:936 within `copyto!' @ simdloop.jl:0
L1136:
        movq    %r10, 32(%rbx)
[…]
