### 2021-12-22 #3965 @mkitti
Similarly, for Python/NumPy there is Numba which also uses LLVM for compilation.

The distinction therefore is not as simple as interpreted versus compiled now. The current dichotomy is that MATLAB, Julia, and Numba are designed for dynamic just-in-time compilation versus static ahead-of-time compilation for C++/Castor, at least in the normal mode of operation. There are definite advantages to static AOT compilation that the authors can focus on.

Frankly, the steps for setting up compilation and the needed environment are not trivial. MATLAB, Julia, and Numba package their dynamic compilers and their dependencies to facilitate their use. Perhaps Castor would benefit by being packaged for an ecosystem such as Conda-Forge. Several C++ header only libraries exist in Conda Forge such as [ensmallen](https://github.com/conda-forge/ensmallen-feedstock). Conda Forge also has [VTK v9.1.0](https://anaconda.org/conda-forge/vtk) and [OpenBlas v0.3.18](https://anaconda.org/conda-forge/openblas) and can be used for Windows, macOS, and Linux.
