### 2021-12-22 #3965 @mkitti
"C++ Development With Conda" is outlined by Antoine Prouvost in the following blog post.
https://www.prouvost.dev/post/2021/c-development-with-conda/
