### 2021-12-22 #3965 @matthieuaussal
@mkitti :
We will gradually modify the installation procedure so that it is more robust on all OS. We will solve the majority of the problems by offering a Conda package. We will try to do it asap.

Indeed, the purpose of the timings is not benchmarking but semantics. Is it really necessary to specify it explicitly in the article, insofar as the matrices involved are of minimal size?

For the rest, we note your comments. Do we need to change the text further?

@jbytecode :
We change "statement of field" to "state of the field".

Do you have any news from the second reviewer?
