### 2021-12-22 #3965 @mkitti
I will be more specific. My specific question is if we can call MATLAB and Julia **interpreted** languages as implied in the following two paragraphs. Python/NumPy is perhaps the only interpreted language out of the three at the moment within its standard execution environment.

Specifically, I do not think that the use of "interpreted" in the following paragraphs is an accurate or true statement in application to MATLAB and Julia.

https://github.com/leprojetcastor/castor/blob/master/paper/paper.md?plain=1#L41

> For a developer accustomed to the Matlab language, it is natural to turn to prototyping tools such as Numpy or Julia, to produce open-source scientific computing codes. Indeed, these **interpreted** languages today offer similar semantics and performance, with well-established user communities.

https://github.com/leprojetcastor/castor/blob/2888b6c53d71a8a8b2fd30fbe6947dd096141828/paper/paper.md?plain=1#L68

> But the use of an **interpreter** limits the portability of the code and the transition from the prototype phase to production use can become difficult. Indeed, even if there are tools to automatically generate C/C++ code from a high-level language (as *Matlab Coder*), this work is still often done manually.

1. What makes MATLAB and Julia distinct from **compiled** languages?
2. MATLAB uses the MATLAB Execution Engine as described below. That seems to involve some kind of compilation according to Mathwork's description, though this is hard for me verify since it is a proprietary product. However, MATLAB code execution is faster than it has been in the past. Can MATLAB still be considered **interpreted**?

> The improved architecture uses just-in-time (JIT) compilation of all MATLAB code with a single execution pathway. The engine offers improved language quality and provides a platform for future enhancements.

3. Julia is normally compiled as I demonstrated above. I can use interactive tools such as `@code_llvm` and `@code_native` to see the LLVM IR and machine assembly code for a Julia function. Why is this being called "interpreted"?

4. There is a [JuliaInterpreter.jl](https://juliadebug.github.io/JuliaInterpreter.jl/stable/). However, it is only used for debugging purposes. The existence of the `Cling` interpreter does not make C++ an interpreted language.

5. How does the use of an **interpreter** limit the portability of code versus compiled languages? Recently the new Apple M1 processor was released. Notably, it was initially released with support for Python but not the compiled language Fortran. [`MicroPython`](https://micropython.org/) is an interpreter that works on small microprocessors. The portability of an interpreted or compiled language depends on the existence of an interpreter or compiler for the desired platform. If an interpreter were written in portable C, I would argue that it's interpreted language is as portable as C.

6. My suggestion is to focus on the virtues of **static compilation** and specifically on the portability of C++ rather than making very general statements about compiled or interpreted languages. The static compilation mode of C++ allows one to discover errors at compile time rather than run time since it is possible to determine the execution of the code completely . Dynamic just-in-time compilation as used by Julia comes with that caveat that execution can only be verified just before execution and that execution can be modified after compilation by the introduction of new methods. A limitation of MATLAB specifically is that its proprietary runtime is only available where Mathworks provides it due its proprietary closed source nature. Be specific about each language and avoid overly broad generalizations.

The above two paragraphs that I provided a link to need to be revised.
