### 2021-12-23 #3965 @jbytecode
Dear @matthieuaussal, @mkitti

After a long conversation, I feel I need to get in the area, and could you please clarify why do you insist that Julia is an interpreted language in the paper? As well as known, Julia is a dynamic and a compiled language with a special multiple dispatch feature. Nowadays, the distinction between the terms 'compiled' and 'interpreted' are a little bit different (Compilation to bytecode, compilation to machine code, compilation from bytecode to machine code, just in time compilation, transpiling, interpretation, etc.). Again, as well as I know, Julia compiles any piece of code before it run it. The compilation process is dependent on the function arguments and the return type. With this feature, Julia is comparable with the compiled languages (such like C, Rust, and C++), but not with Python.

The second thing that should be clarified is, why we need a yet another open source Matlab when Julia is around. I don't mean Castor does not solve any problems or does not bring any ease. But we have to clarify those issues nicely. Of course, comparing things, may include speed or performance tests. And these tests should take care of the precompilation and compilation times - and this issue is also related to hybrid compilation & runtime semantics of languages which I mentioned in the first paragraph.

Sorry for getting down on the field, I hope I've made a contribution.

Thank you!
