### 2021-12-23 #3965 @matthieuaussal
Dear @jbytecode,
Dear @mkitti,

Thank you for the clarification.

In the first paragraph, we removed the keyword "interpreted" so as not to confuse dynamic and static compilation.

In the second paragraph, we removed the elements subject to debate by refocusing the discussion on the need. Indeed, beyond the questions of which language is the best (each one having their own opinion), the fact that there are still many developments in native C ++, and Castor comes to give a serious hand in the prototyping phase. For example, I also do real-time audio signal processing (VST plug-in), the community works in native C ++ (Juce library) and castor really saves development time.

I also modified the first paragraph of statement of need in this sense.

Finally, as we have already mentioned, we do not want to make performance comparisons because that is not what we want to say. As we can already see, the performance comparison is difficult (impossible?) to achieve objectively.
