### 2021-12-23 #3965 @jbytecode
of course, you don't have to perform serious benchmarking unless you don't have any performance claims.

If Castor really reduces the development time, yes this can be considered as a contribution from my side.

I am passing my turn and giving back the rudder to the reviewers.
