### 2021-12-29 #3965 @jbytecode
Dear reviewers @mkitti and @pitsianis

The author(s) made some corrections/additions and the paper has been a little bit changed. Are you agreed with these changes?

Could you please continue to review?

Thank you in advance and have a nice new year! 🎈🙃
