### 2021-12-29 #3965 @mkitti
1. Could the authors clarify how others may make contributions to the software?
2. There are missing references for NumPy and Julia.
3. Can we clarify if we expect the VTK version 8.2 dependency to work via Msys on Windows?
4. On Ubuntu, I am not clear how to install openblas other than `sudo apt install libopenblas-dev`. What version is required?
