### 2021-12-29 #3965 @mkitti
I currently have libopenblas-dev 0.2.20+ds-4 from  https://packages.ubuntu.com/bionic/libopenblas-dev.

I suspect that the authors tested against libopenblas-dev 0.3.8+ds-1 https://packages.ubuntu.com/focal/libopenblas-dev

It looks like the fix is in https://github.com/xianyi/OpenBLAS/pull/1347 so at least version libopenblas-dev 0.3 is required by Castor. Could the authors confirm this is the case?
