### 2021-12-29 #3965 @mkitti
Could the authors specify what the version of all the dependencies they tested again are?

Specify the versions of the following:
1. cmake
2. vtk
3. openblas
4. Ubuntu
5. brew
6. MSYS2
7. freeglut3-dev
8. base-devel
9. mingw-w64-x86_64-gcc
10. mingw-w64-x86_64-cmake
11. mingw-w64-x86_64-openblas
12. mingw-w64-x86_64-vtk
