### 2022-01-03 #3965 @mkitti
Hello again. I hope everyone had a happy new years and holidays. As we emerge from the break, I think the fastest solution I have to verify the software is to spin up some virtual machines.

It would be great to know more about the author's computing environment(s), such as the software versions used, so that I can configure those virtual machines to match theirs.
