### 2022-01-05 ecd3202
Add julia, eigen and numpy
