+++ paper/paper.bib
@@ -1,8 +1,38 @@
+@article{bezanson2012julia,
+  title={Julia: A fast dynamic language for technical computing},
+  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
+  journal={arXiv preprint arXiv:1209.5145},
+  year={2012}
+} 
+
 @article{hackbush:1999,
   author = {{Hackbusch}, W.},
   journal = {Computing, 62(2), 89–108},
   title = {{A sparse matrix arithmetic based on H-matrices. Part 1: Introduction to  H-matrices}},
   year = {1999}
+}
+
+@Article{         harris2020array,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
 @article{sparse:2011,
@@ -62,6 +92,13 @@
   year = {2014},
   journal = {Geneva, Switzerland: International Organization for Standardization},
 }
+
+@misc{eigenweb,
+  author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+  title = {Eigen v3},
+  howpublished = {http://eigen.tuxfamily.org},
+  year = {2010}
+ }
 
 @misc{fembem:21,
   author = {Aussal, M. and Bakry, M.},
