### 2022-01-05 e1f4a8b
add Matlab citation and remove numpy
