+++ paper/paper.bib
@@ -5,6 +5,25 @@
   year={2012}
 } 
 
+@BOOK{blaslapack:00,
+      AUTHOR = {Anderson, E. and Bai, Z. and Bischof, C. and
+                Blackford, S. and Demmel, J. and Dongarra, J. and
+                Du Croz, J. and Greenbaum, A. and Hammarling, S. and
+                McKenney, A. and Sorensen, D.},
+      TITLE = {{LAPACK} Users' Guide},
+      EDITION = {Third},
+      PUBLISHER = {Society for Industrial and Applied Mathematics},
+      YEAR = {1999},
+      ADDRESS = {Philadelphia, PA},
+      ISBN = {0-89871-447-8 (paperback)} }
+
+@article{cpp:14,
+  author = {ISO/IEC},
+  title = {International Standard ISO/IEC 14882:2014(E) – Programming Language C++},
+  year = {2014},
+  journal = {Geneva, Switzerland: International Organization for Standardization},
+}
+
 @article{hackbush:1999,
   author = {{Hackbusch}, W.},
   journal = {Computing, 62(2), 89–108},
@@ -12,28 +31,13 @@
   year = {1999}
 }
 
-@Article{         harris2020array,
- title         = {Array programming with {NumPy}},
- author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
-                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
-                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
-                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
-                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
-                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
-                 Travis E. Oliphant},
- year          = {2020},
- month         = sep,
- journal       = {Nature},
- volume        = {585},
- number        = {7825},
- pages         = {357--362},
- doi           = {10.1038/s41586-020-2649-2},
- publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
-}
+@book{MATLAB:2010,
+year = {2010},
+author = {MATLAB},
+title = {version 7.10.0 (R2010a)},
+publisher = {The MathWorks Inc.},
+address = {Natick, Massachusetts}
+} 
 
 @article{sparse:2011,
  author = {{Davis}, T. A. and {Hu}, Y.},
@@ -71,26 +75,6 @@
   year = {2000},
   doi = {10.1109/38.865875},
   journal = {IEEE Computer graphics and applications, 20(5), 20-27},
-}
-
-@BOOK{blaslapack:00,
-      AUTHOR = {Anderson, E. and Bai, Z. and Bischof, C. and
-                Blackford, S. and Demmel, J. and Dongarra, J. and
-                Du Croz, J. and Greenbaum, A. and Hammarling, S. and
-                McKenney, A. and Sorensen, D.},
-      TITLE = {{LAPACK} Users' Guide},
-      EDITION = {Third},
-      PUBLISHER = {Society for Industrial and Applied Mathematics},
-      YEAR = {1999},
-      ADDRESS = {Philadelphia, PA},
-      ISBN = {0-89871-447-8 (paperback)} }
-
-
-@article{cpp:14,
-  author = {ISO/IEC},
[…]
