### 2022-01-05 04f907d
Add Julia, Matlab and Eigen references
