+++ doc/source/installation.rst
@@ -3,16 +3,33 @@
 Installation
 ============
 
-The simplest way to get the **castor** library is to download the `latest version of header files <https://gitlab.labos.polytechnique.fr/leprojetcastor/castor/-/jobs/artifacts/master/download?job=deploy>`_ and include them during the compilation of your c++ program using the library, see :ref:`label-compilation` in the :ref:`label-basic` section.
+The simplest way to get the dense matrix part of the framework **castor** is to download the `latest version of header files <https://gitlab.labos.polytechnique.fr/leprojetcastor/castor/-/jobs/artifacts/master/download?job=deploy>`_ and include `matrix.hpp` during the compilation of your c++ program using the library, see :ref:`label-compilation` in the :ref:`label-basic` section.
 
-For a complete installation integrating check dependencies and examples compilation, we describe below the procedure for **MacOS** and **Linux** distributions like Ubuntu. For the installation on **Windows**, see the dedicated section :ref:`label-install-windows`.
+For a complete installation integrating check dependencies and examples compilation, we describe below the procedure.
+
+**MacOS** (11 and 12) and **Ubuntu** (20.04)
+++++++++++++++++++++++++++++++++++++++++++++
+
+Installing the dependencies
+---------------------------
+
+The **castor** framework depends on two external dependencies : a BLAS/LAPACK implementation in order to use optimized linear algebra, and VTK for the graphical rendering.
+
+The BLAS/LAPACK implementation which has been tested are :
+- MKL 2020.0.166 : `MKL informations <https://www.intel.com/content/www/us/en/developer/tools/oneapi/onemkl.html>`_, 
+- OpenBlas 0.3.19 : `OpenBLAS informations <https://www.openblas.net/>`_, 
+- vecLib : from accelerate framework in MacOS.
+
+The version of VTK library which has been tested is `9.1.0 <https://vtk.org/download/>`_.
+
+The last tool to perform installation is ``CMake``, at least version `3.18  <https://cmake.org/download/>`_.
+
+Note : on macOS it is recommended to install these dependencies with `brew <https://brew.sh/>_`
 
 From git repository with CMake
-++++++++++++++++++++++++++++++
+------------------------------
 
 You can install the **castor** library from source with ``CMake``. The source files of the library is available here `<https://gitlab.labos.polytechnique.fr/leprojetcastor/castor>`_.
-
-On Linux and macOS platforms :
 
 .. code::
 
@@ -31,163 +48,13 @@ The linear algebra part and the visualization part of the library depend respect
 
     $ cmake -DCMAKE_INSTALL_PREFIX=path/to/install/directory -DCMAKE_PREFIX_PATH="/path/to/optimized/blas;/path/to/vtk/" ..   
 
-Installing the dependencies
-+++++++++++++++++++++++++++
+Xcode project for macOS
+-----------------------
 
-Installing BLAS and LAPACK
-..........................
+To create a Xcode project with ``CMake``, add option `-G Xcode` in the `cmake` command :
+ 
+.. code::
 
-BLAS and LAPACK may be obtained through multiple channels:
+    $ cmake -G Xcode -DCMAKE_INSTALL_PREFIX=path/to/install/directory 
 
- - the first possibility it to download the binaries for `OpenBLAS <https://www.openblas.net/>`_. On Ubuntu, they can be obtained with 
-
-    .. code:: text
-
-        sudo apt install libopenblas-dev
-
-    and on MacOS with
-
-    .. code::
-
-        brew install openblas
-
[…]
