### 2022-01-05 a12fb9e
Update installation.rst
