### 2022-01-07 33ae4ab
Update installation.rst with windows 10 recommandations
