+++ doc/source/installation.rst
@@ -59,3 +59,59 @@ To create a Xcode project with ``CMake``, add option `-G Xcode` in the `cmake` c
     $ cmake -G Xcode -DCMAKE_INSTALL_PREFIX=path/to/install/directory 
 
 This project is created in the `build` directory.
+
+Windows (10)
+++++++++++++
+
+This solution has been tested on Windows 10 only (but may work on other version). Since there is no *built_in* available package manager, the different components will be installed *by-hands* using only *Windows-like* tools.
+
+C++ compiler and CMake
+----------------------
+
+The first step is to install a suitable C++ compiler. In the following instructions we will only use the compiler provided with [Visual Studio](https://visualstudio.microsoft.com/fr/downloads/), version 2019 or later (previous version may work but have not been tested). The Visual Studio framework also provides a customized command prompt named `x64 Native Tools Command Prompt`.
+
+We will also install the [CMake](https://cmake.org) tools. Go to [https://cmake.org](https://cmake.org) and download the latest binary distribution for Windows. After installing CMake, open the `x64 Native Tools Command Prompt` and execute the command `cmake-gui`. If the command fails, find the install folder of CMake and add the `CMakeInstallFolder\bin` subfolder to the Windows `%PATH%` environment variable, restart the command prompt and try again. The graphical interface of CMake should open (close it for now).
+
+BLAS/LAPACK library
+-------------------
+
+The simplest way is probably to install [OpenBLAS](https://www.openblas.net) which implements both interfaces. Compiling the library can quickly become painful as a Fortran compiler is required. Thankfully, the developers have made precompiled binaries available. Installing OpenBLAS can be done following these steps:
+
+1. Go to [https://github.com/xianyi/OpenBLAS/releases](https://github.com/xianyi/OpenBLAS/releases), look for an archive named **OpenBLAS-0.x.x-x64.zip** (or **OpenBLAS-0.x.x-x86.zip** for older architectures) in the section **Assets** corresponding to the version you wish to use and download it.<br>
+**Remark** The demos were tested originally with `OpenBLAS 0.3.12` so any later version should be fine.
+2. Extract the downloaded archive in a folder of your choice (for example, create a folder `openblas`). This folder should now contain three subdirectories:
+    - `openblas\bin` should contain a file named `libopenblas.dll`.
+    - `openblas\include` should contain the header files, including `cblas.h`.
+    - `openblas\lib` should contain `.lib` files (including `libopenblas.lib`) and a `openblas\lib\cmake` subfolder.
+3. Add the subfolder `openblas\bin` to Windows environment variable `%PATH%`.
+
+The BLAS and LAPACK are now ready to use.
+
+**Remark** It is also possible to download the Intel MKL library through the framework `oneAPI/MKL` (see [https://github.com/oneapi-src/oneMKL](https://github.com/oneapi-src/oneMKL) or the [ Intel website](https://www.intel.com/content/www/us/en/developer/tools/oneapi/overview.html)). However, this implementation features different header names and requires a modification of the source files of **castor** (namely replace `cblas.h` by `mkl_cblas.h` wherever it appears). For this reason, we do not insist further.
+
+VTK framework
+-------------
+
+Unfortunately, the developers of the VTK framework do not provide *ready-to-use* binaries meaning that we must compile the sources by ourselves. It is performed as follows:
+
+1. Download the sources of VTK on the main website [https://vtk.org](https://vtk.org). Choose a version of the `9.x.x` branch. Uncompress the archive in a folder of your choice.
+2. Open the `x64 Native Tools Command Prompt` and move to the newly created VTK folder (use the `dir pathToFolder` command). Create a *build* folder using `mkdir build` and move to this folder.
[…]
