### 2022-01-07 #3965 @matthieuaussal
@mkitti
Following your recommendations, we have added references to Matlab, Julia and Eigen in the bibliography of the paper. On the other hand, we did not specify a version of the tools in the paper, because it is intended to be timeless.

However, we have also updated the installation procedure on our documentation, accessible here:

https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/installation.html

You will find the procedure for each OS, this time with the functional versions of each tool and the links to download them. We have verified that everything works on each OS, if we follow the installation procedures proposed by openBlas and VTK.

To test, the easiest would be for you to make an ubuntu virtual machine.

@jbytecode
Could you hear from the second review please?
