### 2022-01-07 #3965 @jbytecode
Dear @matthieuaussal

Thank you for the updates and changes.

Since the review process is currently active, there is no need to ping anyone now, the reviewers can decide to have an action or not while the other reviewer is in action. It seems the second reviewer is also active as they have completed some tasks.

By the way, we can ask them if they capable of having notifications and ask them to update their status.

______________________

Dear @pitsianis,

Could you please update your status and let us know how is your review going?

Thank you in advance.
