### 2022-01-07 #3965 @mkitti
Thank you for the revisions to the paper and installation instructions.

I do not see how indicating the version of the tools that you used would make the software any less timeless. I will defer to the editor whether it is sufficient to have those versions in the installation instructions and not the main text of the paper. However, under most publication procedures a snapshot of those installation instructions as well as all other electronic materials would be included in an archive with the publication, and that would be satisfactory to me.

Minor comments: Some of the link formats look strange in the current HTML rendering of the instructions as the brackets are visible. Also the link for brew is not a hyper link.

On Windows, you may want to see if https://github.com/microsoft/vcpkg may be useful in enhancing your instructions.
