### 2022-01-07 #3965 @jbytecode
@mkitti - It is sufficient to have directives and installation instructions in Github or/and documentation.
