### 2022-01-07 #3965 @mkitti
That's fine with me. I'm just noticing that the Github repository is out of sync with the website:
https://github.com/leprojetcastor/castor/blob/master/doc/source/installation.rst
https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/installation.html

Do we usually require the use of Zenodo to capture a DOI tagged snapshot of the Git repository?
