### 2022-01-07 #3965 @jbytecode
The host repo should include any installation instructions, however, external links are okay as soon as they are quareeted to be exist.
