+++ doc/source/installation.rst
@@ -25,7 +25,7 @@ The version of VTK library which has been tested is `9.1.0 <https://vtk.org/down
 
 The last tool to perform installation is ``CMake``, at least version `3.18  <https://cmake.org/download/>`_.
 
-Note : on macOS it is recommended to install these dependencies with `brew <https://brew.sh/>_`
+**Note** : on macOS it is recommended to install these dependencies with `brew <https://brew.sh/>`_.
 
 From git repository with CMake
 ------------------------------
@@ -68,17 +68,16 @@ This solution has been tested on Windows 10 only (but may work on other version)
 C++ compiler and CMake
 ----------------------
 
-The first step is to install a suitable C++ compiler. In the following instructions we will only use the compiler provided with [Visual Studio](https://visualstudio.microsoft.com/fr/downloads/), version 2019 or later (previous version may work but have not been tested). The Visual Studio framework also provides a customized command prompt named `x64 Native Tools Command Prompt`.
+The first step is to install a suitable C++ compiler. In the following instructions we will only use the compiler provided with `Visual Studio <https://visualstudio.microsoft.com/fr/downloads/>`_, version 2019 or later (previous version may work but have not been tested). The Visual Studio framework also provides a customized command prompt named `x64 Native Tools Command Prompt`.
 
-We will also install the [CMake](https://cmake.org) tools. Go to [https://cmake.org](https://cmake.org) and download the latest binary distribution for Windows. After installing CMake, open the `x64 Native Tools Command Prompt` and execute the command `cmake-gui`. If the command fails, find the install folder of CMake and add the `CMakeInstallFolder\bin` subfolder to the Windows `%PATH%` environment variable, restart the command prompt and try again. The graphical interface of CMake should open (close it for now).
+We will also install the `CMake <https://cmake.org>`_ tools. Download the latest binary distribution for Windows. After installing CMake, open the `x64 Native Tools Command Prompt` and execute the command `cmake-gui`. If the command fails, find the install folder of CMake and add the `CMakeInstallFolder\bin` subfolder to the Windows `%PATH%` environment variable, restart the command prompt and try again. The graphical interface of CMake should open (close it for now).
 
 BLAS/LAPACK library
 -------------------
 
-The simplest way is probably to install [OpenBLAS](https://www.openblas.net) which implements both interfaces. Compiling the library can quickly become painful as a Fortran compiler is required. Thankfully, the developers have made precompiled binaries available. Installing OpenBLAS can be done following these steps:
+The simplest way is probably to install `OpenBLAS <https://www.openblas.net>`_ which implements both interfaces. Compiling the library can quickly become painful as a Fortran compiler is required. Thankfully, the developers have made precompiled binaries available. Installing OpenBLAS can be done following these steps:
 
-1. Go to [https://github.com/xianyi/OpenBLAS/releases](https://github.com/xianyi/OpenBLAS/releases), look for an archive named **OpenBLAS-0.x.x-x64.zip** (or **OpenBLAS-0.x.x-x86.zip** for older architectures) in the section **Assets** corresponding to the version you wish to use and download it.<br>
-**Remark** The demos were tested originally with `OpenBLAS 0.3.12` so any later version should be fine.
+1. Go to `https://github.com/xianyi/OpenBLAS/releases <https://github.com/xianyi/OpenBLAS/releases>`_, look for an archive named **OpenBLAS-0.x.x-x64.zip** (or **OpenBLAS-0.x.x-x86.zip** for older architectures) in the section **Assets** corresponding to the version you wish to use and download it. The demos were tested originally with `OpenBLAS 0.3.12` so any later version should be fine.
[…]
