### 2022-01-07 #3965 @matthieuaussal
@mkitti
I fixed hyperlink in the windows part of the documentation.

@jbytecode
The online documentation with continuous integration is now hosted on the gitlab server of our institution, Ecole Polytechnique, which is sustainable over time. Links to this gitlab have been given on the Github page. As a full migration to github requires a lot of work, we plan to do so but we don't know when.
