### 2022-01-07 #3965 @pitsianis
@matthieuaussal, in many math papers, the word MATLAB is typeset as `\textsc{Matlab}`
