### 2022-01-11 8110664
add low-rank builder, hfull2 and dimensions check with inline function
