+++ include/castor/hmatrix.hpp
@@ -199,6 +199,7 @@ public:
     // CONSTRUCTOR
     hmatrix(){};
     hmatrix(std::size_t m, std::size_t n, double tol, T v=0);
+    hmatrix(matrix<T>const& A, matrix<T>const& B, double tol);
     template<typename S>
     hmatrix(matrix<S>const& X, matrix<S>const& Y, double tol, matrix<T>const& M);
     template<typename S>
@@ -215,6 +216,8 @@ public:
     void full2lowrank();
     void fusion();
     void hfull(matrix<T>& M, matrix<std::size_t> I={}, matrix<std::size_t> J={}) const;
+    void hfull2(matrix<T>& M, matrix<std::size_t> idx, matrix<std::size_t> jdx,
+              matrix<std::size_t> I={}, matrix<std::size_t> J={}) const;
     void hinv();
     void hllowsolve(hmatrix<T>const& Lh);
     void hllowsolve(matrix<T>& B, matrix<std::size_t> I={}) const;
@@ -238,6 +241,7 @@ public:
     hmatrix<T>& operator+=(T b);
     hmatrix<T>& operator*=(T b);
     hmatrix<T>& operator+=(hmatrix<T>const& Bh);
+    matrix<T>   operator()(matrix<std::size_t>const& I, matrix<std::size_t>const& J) const;
     
     // ACCESSORS
     hmatrix<T>const&          getChildren(int i) const {return m_chd[i];};
@@ -280,6 +284,19 @@ hmatrix<T>::hmatrix(std::size_t m, std::size_t n, double tol, T v)
     m_tol    = tol;
     m_dat[0] = matrix<T>(m,1,v);
     m_dat[1] = matrix<T>(1,n,1.);
+}
+
+//==========================================================================
+// [.hmatrix]
+/// Low-rank constructor, fill leaf with tensor product. No tree is needed.
+template<typename T>
+hmatrix<T>::hmatrix(matrix<T>const& A, matrix<T>const& B, double tol)
+{
+    m_typ    = 1;
+    m_siz    = {size(A,1),size(B,2)};
+    m_tol    = tol;
+    m_dat[0] = A;
+    m_dat[1] = B;
 }
 
 //==========================================================================
@@ -543,6 +560,51 @@ void hmatrix<T>::hfull(matrix<T>& M, matrix<std::size_t> I, matrix<std::size_t> 
 }
 
 //==========================================================================
+// [hmatrix.full2]
+/// In-place conversion to dense sub-matrix :
+/// Ah(I,J) -> M.
+template<typename T>
+void hmatrix<T>::hfull2(matrix<T>& M, matrix<std::size_t> idx, matrix<std::size_t> jdx,
+                      matrix<std::size_t> I, matrix<std::size_t> J) const
+{
+    if (isempty(I)) {I = range(0,size(M,1));}
+    if (isempty(J)) {J = range(0,size(M,2));}
+    if (numel(I)!=numel(idx) || numel(J)!=numel(jdx))
+    {
+        error(__FILE__, __LINE__, __FUNCTION__,"Dimensions must agree.");
+    }
+    if (m_typ==0)
+    {
+        for (int h=0; h<4; ++h)
+        {
+            matrix<std::size_t> idxh, jdxh, Ih, Jh, Itmp, Jtmp;
+            std::tie(idxh,Itmp) = argintersect(m_row[h],idx);
+            std::tie(jdxh,Jtmp) = argintersect(m_col[h],jdx);
[…]
