### 2022-01-11 #3965 @mkitti
With versions as specified by the authors at https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/installation.html I have been able to compile and execute the code on Ubuntu 20.04 (LTS).

Some comments on this:
1. The authors specify Ubuntu (20.04) in the instructions and a minimum version of cmake 3.18. This is enforced by CMakeLists.txt. However, the [version of cmake available in Ubuntu 20.04 is 3.16](https://packages.ubuntu.com/focal/cmake). Ubuntu 21.04 is the earliest version that includes [cmake 3.18](https://packages.ubuntu.com/hirsute/cmake), so perhaps 21.04 should be recommended. `cmake` is a pretty common and basic tool, so it would be nice if one could just the packaged version for that.
2. The authors no longer provide installation instructions for the dependencies.
3. I was able to create the build easily by the following.
```
conda -n castor cmake openblas vtk
conda activate castor
mkdir build
cd build
mkdir install
cmake -DCMAKE_INSTALL_PREFIX=./install ..
```
4. The above installed cmake 3.21.3, openblas 0.3.18, vtk 9.1.
