### 2022-01-11 #3965 @mkitti
The reference for NumPy was added and then removed.
