### 2022-01-11 #3965 @mkitti
>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I am unable to identify community guidelines to contribute to the software, report issues, or seek support in the documentation.
