### 2022-01-11 #3965 @mkitti
The examples in the paper make reference to a `castor/linsolve.hpp`. This file does not exist. Perhaps this should be `linalg.hpp`?
