### 2022-01-11 #3965 @mkitti
With demo_graphics, I am not getting past "DRAWNOW". If I close the opened figures, then I get a segmentation fault. I do not get to the "IMAGE WRITE" or "MOVIE WRITE" sections.
