### 2022-01-11 #3965 @mkitti
I also get a segmentation fault when rendering the Figure 1 example, when I close the window.
