### 2022-01-11 #3965 @mkitti
Are there any build instructions for Figure 2. I assume it is from this repository:
https://gitlab.labos.polytechnique.fr/leprojetcastor/fembem

It would be nice to have instructions on how to construct a CMakeLists.txt file to compile the examples in the documentation and the paper.
