### 2022-01-11 #3965 @mkitti
Here is the backtrace via gdb from the segmentation fault that happens when I close the window.
```
Thread 1 "review" received signal SIGSEGV, Segmentation fault.
0x00007ffff57f72dc in _XEventsQueued () from ~/mambaforge-pypy3/envs/castor/lib/./libX11.so.6
(gdb) bt
#0  0x00007ffff57f72dc in _XEventsQueued () from ~/mambaforge-pypy3/envs/castor/lib/./libX11.so.6
#1  0x00007ffff57e81b2 in XPending () from ~/mambaforge-pypy3/envs/castor/lib/./libX11.so.6
#2  0x00007ffff5904e72 in vtkXRenderWindowInteractor::StartEventLoop() () from ~/mambaforge-pypy3/envs/castor/lib/./libvtkRenderingUI-9.1.so.1
#3  0x000055555555866e in castor::figure::drawnow() ()
#4  0x000055555555a34d in castor::drawnow(castor::figure&) ()
#5  0x000055555555983a in main ()
```

The code is just as follows from Figure 1.
```c++
#include <iostream>
#include "castor/matrix.hpp"
#include "castor/graphics.hpp"

using namespace castor;

int main(int argc, char* argv[])
{
    matrix<> X = linspace(0, 10, 100);
    figure fig;
    plot(fig, X, sin(X), {"r-+"}, {"sin(x)"});
    plot(fig, X, cos(X), {"bx"}, {"cox(s)"});
    drawnow(fig);
    return 0;
}
```

![image](https://user-images.githubusercontent.com/8062771/148972476-7b2a2f8d-892b-4ef8-a87e-2bebf3d34cbc.png)
