### 2022-01-11 #3965 @mkitti
The backtrace is similar for demo_graphics
