### 2022-01-13 daa2b8a
add kalman interface for hmatrix
