+++ include/castor/hmatrix.hpp
@@ -211,6 +211,8 @@ public:
     hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, double tol);
     template<typename S>
     hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, double tol, smatrix<T>const& Ms);
+    template<typename S>
+    hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, smatrix<T>const& Ml, hmatrix<T>const& Mh, smatrix<T>const& Mr);
     
     // FUNCTIONS
     void full2lowrank();
@@ -457,6 +459,120 @@ hmatrix<T>::hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, double tol, smatri
     }
 }
 
+//==========================================================================
+// [.hmatrix]
+/// Projector constructor : Ml * Mh * Mr
+template<typename T>
+template<typename S>
+hmatrix<T>::hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, smatrix<T>const& Ml, hmatrix<T>const& Mh, smatrix<T>const& Mr)
+{
+    // Dimensions compatibility
+    if (size(Xb.crd(),1)!=size(Ml,1) || size(Yb.crd(),1)!=size(Mr,2) ||
+        size(Ml,2)!=size(Mh,1) || size(Mh,2)!=size(Mr,1))
+    {
+        error(__FILE__, __LINE__, __FUNCTION__,"Dimensions must agree.");
+    }
+    
+    // Global Indices
+    m_siz = {size(Xb.crd(),1),size(Yb.crd(),1)};
+    m_tol = Mh.getTolerance();
+
+    // Empty leaf
+    if (nnz(Ml)==0 || nnz(Mr)==0)
+    {
+        m_typ       = 1;
+        m_dat[0]    = zeros<T>(m_siz(0),1);
+        m_dat[1]    = zeros<T>(1,m_siz(1));
+    }
+    // Far leaf
+    else if (Mh.getType()==1)
+    {
+        m_typ    = 1;
+        m_dat[0] = mtimes(Ml,Mh.getData(0));
+        m_dat[1] = mtimes(Mh.getData(1),Mr);
+    }
+    // Full leaves
+    else if (Mh.getType()==2 && (Xb.isleaf() || Yb.isleaf()))
+    {
+        m_typ    = 2;
+        m_dat[0] = mtimes(mtimes(Ml,Mh.getData(0)),Mr);
+    }
+    // Full leaf from Mh
+    //   M = (A1) (B1) (C1 C2)
+    //       (A2)
+    else if (Mh.getType()==2)
+    {
+        // Indices of sub matrices
+        matrix<std::size_t> idx, jdx, tmp;
+        std::tie(tmp,jdx) = ind2sub(size(Ml),index(Ml));
+        std::tie(idx,tmp) = ind2sub(size(Mr),index(Mr));
+        idx = unique(idx);
+        jdx = unique(jdx);
+
+        // Sub matrices
+        smatrix<T> A = eval(Ml(row(Ml),jdx));
+        smatrix<T> B = sparse(eval(Mh.getData(0)(jdx,idx)));
+        smatrix<T> C = eval(Mr(idx,col(Mr)));
+
+        // Product
+        smatrix<T> Ms = mtimes(mtimes(A,B),C);
+        (*this) = hmatrix<double>(Xb,Yb,m_tol,Ms);
+    }
+    // Full leaf from tree
+    //   M = (A1 A2) (B1 B2) (C1)
+    //               (B3 B4) (C2)
+    else if (Xb.isleaf() || Yb.isleaf())
+    {
+        // Indices of sub matrices
[…]
