+++ include/castor/hmatrix.hpp
@@ -1991,196 +1991,3 @@ inline hmatrix<T> transpose(hmatrix<T>const& Ah)
 
 }
 
-
-
-
-
-
-
-////==========================================================================
-//// [hmatrix.full2lowrank]
-///// Convert full leaf to low-rank using ACA compression.
-//template<typename T>
-//void hmatrix<T>::full2lowrank()
-//{
-//    if (m_typ==2 && m_tol>0)
-//    {
-//        if (rank(m_dat[0])<min(m_siz)*0.9)
-//        {
-//            m_typ = 1;
-//            std::tie(m_dat[0],m_dat[1]) = qrsvd(m_dat[0],eye(m_siz(1)),m_tol);
-//        }
-//    }
-//}
-
-
-////==========================================================================
-//// [hmatrix.full2lowrank]
-///// Convert full leaf to low-rank using ACA compression.
-//template<typename T>
-//void hmatrix<T>::full2lowrank()
-//{
-//    if (m_typ==2 && m_tol>0)
-//    {
-//        matrix<T> A, B;
-//        bool flag;
-//        std::tie(A,B,flag) = aca(m_dat[0],m_tol);
-//        if (flag)
-//        {
-//            m_dat[0] = A;
-//            m_dat[1] = B;
-//            m_typ    = 1;
-//        }
-//    }
-//}
-
-
-
-
-
-
-
-////==========================================================================
-//// [.hmatrix]
-///// Projector constructor : Ml * Mh * Mr
-//template<typename T>
-//template<typename S>
-//hmatrix<T>::hmatrix(bintree<S>const& Xb, bintree<S>const& Yb, smatrix<T>const& Ml, hmatrix<T>const& Mh, smatrix<T>const& Mr)
-//{
-//
-//    // Dimensions compatibility
-//    if (size(Xb.crd(),1)!=size(Ml,1) || size(Yb.crd(),1)!=size(Mr,2) ||
-//        size(Ml,2)!=size(Mh,1) || size(Mh,2)!=size(Mr,1))
-//    {
-//        error(__FILE__, __LINE__, __FUNCTION__,"Dimensions must agree.");
-//    }
-//
-//    // Global Indices
-//    m_siz = {size(Xb.crd(),1),size(Yb.crd(),1)};
-//    m_tol = Mh.getTolerance();
-//
-//    // Low-rank
-//    if (Mh.getType()==1)
-//    {
-//        m_typ    = 1;
-//        m_dat[0] = mtimes(Ml,Mh.getData(0));
-//        m_dat[1] = mtimes(Mh.getData(1),Mr);
-////        std::tie(m_dat[0],m_dat[1]) = qrsvd(m_dat[0],m_dat[1],m_tol);
-//    }
-//    // Empty leaves
-//    else if (nnz(Ml)==0 || nnz(Mr)==0)
-//    {
-//        m_typ       = 1;
-//        m_dat[0]    = zeros<T>(m_siz(0),1);
-//        m_dat[1]    = zeros<T>(1,m_siz(1));
-//        m_dat[0](0) = M_EPS(T);
-//        m_dat[1](0) = M_EPS(T);
-//    }
-//    // Far leaves
-//    else if (Xb.box().isfar(Yb.box()))
-//    {
-//        // Indices of sub matrices
-//        matrix<std::size_t> idx, jdx, tmp;
-//        std::tie(tmp,jdx) = ind2sub(size(Ml),index(Ml));
-//        std::tie(idx,tmp) = ind2sub(size(Mr),index(Mr));
-//        idx = unique(idx);
-//        jdx = unique(jdx);
-//
-//        // Sub h-matrix with ACA compression
[…]
