### 2022-01-14 #3965 @matthieuaussal
> Le 11 janv. 2022 à 16:40, Mark Kittisopikul ***@***.***> a écrit :
>
>
> The backtrace is similar for demo_graphics
>
