### 2022-01-14 #3965 @matthieuaussal
@pitsianis
I wrote matlab with $\textsc{Matlab}$, but that doesn't seem to change

@mkitti
1. The reference to numpy has been removed because we never mention numpy in the text. We only kept the references of the languages ​​that are compared in the article. Otherwise, it would also be necessary to cite all equivalent software, which cannot be exhaustive.

2. There is no community guideline as we are unable to commit to long term support. Today, we are giving our tool as is, the LGPL license allowing everyone to do what they want with it.

3. We refer to "linsolve" function included in linalg.hpp, wich is not sor far from "backslash", and equivalent to the "linsolve" function of Matlab.

4. We have this segmentation error on windows and linux since we upgraded to VTK 9.1 (not macos). We have not yet identified the source properly, but it seems to be due to the pointers offered by VTK, which do not empty when you close the windows. We are waiting for the next VTK update before acting if the bug persists.

5. In the demo-graphics.cpp example, the reading / writing of images and videos is after the figures, because again we encounter errors under mac and windows since vtk 9.1. We are still waiting for the next update of VTK.
