### 2022-01-14 #3965 @mkitti
![Screenshot_20220114-091241_GitHub.jpg](https://user-images.githubusercontent.com/8062771/149529036-55fab2e5-fd9f-47d1-9d48-3777161a2bdd.jpg)

The linsolve.hpp does not exist yet it is used in the examples.
