### 2022-01-14 fe64f01
fix linalg.hpp instead of linsolve
