### 2022-01-14 #3965 @matthieuaussal
ok, thanks. I fixed it

> Le 14 janv. 2022 à 15:13, Mark Kittisopikul ***@***.***> a écrit :
>
>
>  <https://user-images.githubusercontent.com/8062771/149529036-55fab2e5-fd9f-47d1-9d48-3777161a2bdd.jpg>
> The linsolve.hpp does not exist yet it is used in the examples.
>
