### 2022-01-14 #3965 @jbytecode
@matthieuaussal the software repo should include community guidelines as mentioned in reviewer tasks as:

- Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
