### 2022-01-14 #3965 @matthieuaussal
> Le 14 janv. 2022 à 15:17, Matthieu Aussal ***@***.***> a écrit :
>
> ok, thanks. I fixed it
>
>
>> Le 14 janv. 2022 à 15:13, Mark Kittisopikul ***@***.*** ***@***.***>> a écrit :
>>
>>
>>  <https://user-images.githubusercontent.com/8062771/149529036-55fab2e5-fd9f-47d1-9d48-3777161a2bdd.jpg>
>> The linsolve.hpp does not exist yet it is used in the examples.
>>
>> —
>> Reply to this email directly, view it on GitHub <https://github.com/openjournals/joss-reviews/issues/3965#issuecomment-1013151801>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AHFHRGGHQCBEIWIUYOTJCCLUWAVRTANCNFSM5JDT4MGQ>.
>> Triage notifications on the go with GitHub Mobile for iOS <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>> You are receiving this because you were mentioned.
>>
>
