### 2022-01-14 #3965 @matthieuaussal
ok, I will see with my colleagues to agree on a clear and precise policy.

> Le 14 janv. 2022 à 15:18, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> @matthieuaussal <https://github.com/matthieuaussal> the software repo should include community guidelines as mentioned in reviewer tasks as:
>
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
