### 2022-01-14 #3965 @mkitti
Also the Julia line `(M .- 1) .* Matrix(I, size(M,1), size(M,2))` could be shorter: `(M .- 1) .* Matrix(I, size(M))`.
