### 2022-01-14 #3965 @mkitti
Try $\textsc{MATLAB}$. sc indicates small capital letters
