### 2022-01-15 #3965 @mkitti
Regarding VTK, I understand that an upgrade to 9.1 is related to the segmentation fault issues. What is the latest version of VTK that does not have the segmentation fault issues?

It seems strange to state "The version of VTK library which has been tested" and not state are known issues with that testing.
