### 2022-01-24 #3965 @mkitti
Correct. I have a few outstanding questions above.

It would be nice to clarify the segmentation fault or have some note that it is currently expected with VTK 9.1 on some operating systems.
