### 2022-01-24 #3965 @jbytecode
@pitsianis whedon successfully compiles the pdf, as in http://res.cloudinary.com/hkvhhndao/image/upload/v1643040300/wkkkql3ryhyiejpyqona.pdf

and I will try it again here
