### 2022-01-24 #3965 @mkitti
> It would be nice to clarify the segmentation fault or have some note that it is currently expected with VTK 9.1 on some operating systems.

Let me be more specific:
1. Is there a version of VTK that does not result in a segmentation fault on Windows and Linux? Does 9.0 work?
2. Can we do a `git bisect` on https://github.com/Kitware/VTK to narrow down the causing commit?
3. Can we point an bug report or issue with VTK on this issue? Perhaps this already describes the issue?
https://gitlab.kitware.com/vtk/vtk/-/issues/18372
