+++ CONTRIBUTING.md
@@ -1,38 +1,28 @@
-# Contributing Code
+# How to contribute
 
-## How to contribute
 The preferred way to contribute to Castor is to fork the [main repository](https://github.com/leprojetcastor/castor) on Github.
 
-1. *Search or open an [issue](https://github.com/UCD4IDS/WaveletsExt.jl/issues) for this project*:   
+1. *Search or open an [issue](https://github.com/leprojetcastor/castor/issues) for this project*:   
     * Search for and click into a related issue that you would like to contribute to. If the issue has not been assigned to anyone, leave a comment and assign the issue to yourself by clicking on the 'Assignees' button on the right.
 
     * If no such issues have been raised, open a new issue by clicking on the 'New issue' on the top right corner. Write a brief description of the issue and wait for the feedbacks from core developers before moving on to the next steps.
 
-1. *Fork the [project repository](https://github.com/UCD4IDS/WaveletsExt.jl)*: Click on the 'Fork' button on the top right corner of the page. This creates a copy of the code under your account on the GitHub server. 
+2. *Fork the [project repository](https://github.com/leprojetcastor/castor)*: Click on the 'Fork' button on the top right corner of the page. This creates a copy of the code under your account on the GitHub server. 
 
-1. *Clone this copy to your local disk*: Open up a terminal on your computer, navigate to your preferred directory, and copy the following.
+3. *Clone this copy to your local disk*: Open up a terminal on your computer, navigate to your preferred directory, and copy the following.
 ```
-$ git clone git@github.com:<YourGithubUsername>/WaveletsExt.jl.git
-$ cd WaveletsExt
-```
-
-4. *Instantiate the project*: Open up your Julia REPL, and instantiate the current project in the package manager.
-```
-$ julia
+$ git clone git@github.com:<YourGithubUsername>/castor.git
+$ cd castor
 ```
 
-```julia
-julia> ]
-(v1.7) pkg> activate .
-(WaveletsExt) pkg> instantiate
-```
+4. *Instantiate the project*: Install Castor and dependencies and compile the demos following the [user guide](https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/installation.html).
 
 5. *Create a branch to hold your changes*:
 ```
 git checkout -b my-feature
 ```
 
-6. Work on this copy on your computer using your preferred code editor such as VSCode. Make sure you add the [corresponding tests](https://docs.julialang.org/en/v1/stdlib/Test/) in the `test/` directory where appropriate. Use Git to do the version control. When you're done editing, do the following to record your changes in Git:
+6. Work on this copy on your computer using your preferred code editor such as Xcode, Visual Studio, VSCode, etc. Make sure you add the [corresponding tests] in the appropriate `demo_xxx.cpp` file. Use Git to do the version control. When you're done editing, do the following to record your changes in Git:
 ```
 $ git add modified_files
 $ git commit
@@ -43,46 +33,7 @@ $ git commit
 $ git push -u origin my-feature
 ```
 
-8. Finally, go to the web page of your fork of the WaveletsExt.jl repository, and *click 'Pull request'* to send your changes to the maintainers for review. This will send an email to the committers.
+8. Finally, go to the web page of your fork of the castor repository, and *click 'Pull request'* to send your changes to the maintainers for review. This will send an email to the committers.
 
-(If any of the above seems like magic to you, then look up the [Git documentation](https://git-scm.com/doc) on the web.)
+If any of the above seems like magic to you, then look up the [Git documentation](https://git-scm.com/doc) on the web. It is recommended to check that your contribution complies with the following rules before submitting a pull request: All public methods should have informative docstrings with sample usage presented, compatible with the documentation process.
 
-### Feature contribution notes
[…]
