### 2022-01-25 842a8d4
Update graphics.rst
