### 2022-01-25 #3965 @matthieuaussal
@mkitti
We add a community guidelines at the root of the castor repertory :
--> https://github.com/leprojetcastor/castor/blob/master/CONTRIBUTING.md

We can't back to VTK8.2 because it's not the same API (no retro-compatibility). To fix it, we have to rewrite the major part of graphics.hpp, because the segfault is due to a pointers clearing inside the VTK function :

--> vtkContextView::GetInteractor()::Start();

So, waiting we get time to rewrite graphics.hpp and following your advise, we add a note in the documentation to alert linux/windows users, and invit them to use the drawnow() function at the end of them programs:

--> https://leprojetcastor.gitlab.labos.polytechnique.fr/castor/graphics.html

With this precaution, everything works fine.
