### 2022-01-27 #3965 @pitsianis
@matthieuaussal  here are some trivial changes that give the draft a consistent look.

Use highlighting in all codes not just some.

`C++`  in some places is in capitals, in some it has a space etc.

Capitalize correctly the reference titles, for example "university of florida"

File names appear some of the times in italics and in some straight. Use a fixed width style.

The first code example is rather meaningless. May I suggest you define a (sparse) symmetric  0-1 matrix `A` denoting the adjacency matrix of an undirected graph and use it to calculate the ____MATLAB____ expression `sum(A .* (A * A),2) / 2` a vector which reveals the number of triangles at each graph vertex.
