### 2022-02-01 #3965 @pitsianis
Do not use an itemized list in multiline formulae (Application with a FEM/BEM simulation).

The example FEM/BEM simulation used here is very effective in demonstrating `Castor`, however, I had a hard time mapping what is what and moving back and forth between pages 8, 9 and 10, the mathematical description and the code implementation. Wouldn't be better to interleave the math with the code, as if in a Jupyter notebook? Match the math equations with the corresponding code, and show the listing in parts. Place the complete code example in the repository.

These should cover all my concerns with the paper, and I will focus on the code next and wrap up my review.
