+++ paper/paper.bib
@@ -10,7 +10,7 @@
                 Blackford, S. and Demmel, J. and Dongarra, J. and
                 Du Croz, J. and Greenbaum, A. and Hammarling, S. and
                 McKenney, A. and Sorensen, D.},
-      TITLE = {{LAPACK} Users' Guide},
+      TITLE = {{LAPACK} Users' guide},
       EDITION = {Third},
       PUBLISHER = {Society for Industrial and Applied Mathematics},
       YEAR = {1999},
@@ -19,7 +19,7 @@
 
 @article{cpp:14,
   author = {ISO/IEC},
-  title = {International Standard ISO/IEC 14882:2014(E) – Programming Language C++},
+  title = {International standard ISO/IEC 14882:2014(E) – Programming language C++},
   year = {2014},
   journal = {Geneva, Switzerland: International Organization for Standardization},
 }
@@ -41,7 +41,7 @@ address = {Natick, Massachusetts}
 
 @article{sparse:2011,
  author = {{Davis}, T. A. and {Hu}, Y.},
- title = {The University of Florida sparse matrix collection},
+ title = {The university of florida sparse matrix collection},
  journal = {ACM Transactions on Mathematical Software (TOMS), 38(1), 1-25},
  doi = {10.1145/2049662.2049663},
  year = 2016
