### 2022-02-18 #3965 @matthieuaussal
> Le 18 févr. 2022 à 09:49, whedon ***@***.***> a écrit :
>
>
> PDF failed to compile for issue #3965 <https://github.com/openjournals/joss-reviews/issues/3965> with the following error:
>
>  Error producing PDF.
> ! Missing \endgroup inserted.
> <inserted text>
>                 \endgroup
> l.819 \begin{eqnarray}
>
> Looks like we failed to compile the PDF
>
