### 2022-03-06 #3965 @pitsianis
A few minor observations for the document manuscript as I read through to complete my review:

matlab -> Matlab in the title

Matrix(I,size(M,1),size(M,2)) -> Matrix(I,size(M))

Indent the matrix entries properly for `Matlab`, `Julia` and `C++` in all listings

```
M = [ 1  2  3;
      4  5  6;
      7  8  9;
     10 11 12];

M « 1,  2,  3,
    4,  5,  6,
    7,  8,  9,
   10, 11, 12;

matrix<>M =
{{ 1,  2,  3},
 { 4,  5,  6},
 { 7,  8,  9},
 {10, 11, 12}};
```

Refer to Castor in the same way, please.

`argsort` is a function from `NumPy` which is not properly covered in this article, perhaps you should use `sortperm` from `Julia` as well?

Function names such as sgemm should be in teletype font, like `sgemm`

Citation for sparse matrices should acknowledge the datatype not just the exceptional collection ;-) Perhaps Tewarson's Sparse Matrices 1973 book?
