### 2022-03-10 #3965 @matthieuaussal
> Le 7 mars 2022 à 17:26, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> Dear @matthieuaussal <https://github.com/matthieuaussal>
> It seems our reviewers have just completed their major tasks. Please consider applying the latest minor suggestions, then I am going to start my editorial stuff.
>
> Thank you in advance.
>
