+++ paper/paper.bib
@@ -39,12 +39,11 @@ publisher = {The MathWorks Inc.},
 address = {Natick, Massachusetts}
 } 
 
-@article{sparse:2011,
- author = {{Davis}, T. A. and {Hu}, Y.},
- title = {The university of florida sparse matrix collection},
- journal = {ACM Transactions on Mathematical Software (TOMS), 38(1), 1-25},
- doi = {10.1145/2049662.2049663},
- year = 2016
+@article{sparse:1973,
+ author = {{Tewarson}, R. P.},
+ title = {Sparse matrices},
+ journal = {New York: Academic press, Vol. 69},
+ year = 1973
 }
 
 @article{rjasanow:2002,
