### 2022-03-10 #3965 @matthieuaussal
Dear @pitsianis,

We apply all your recommandations, except teletype font for `sgemm`, as it was already done (if we well understand).  We don't cover `argsort` in this article, but it's fully done in the documentation. `Sortperm` function would be to achieve  in furthers developments.

Many thanks for your review
