+++ paper/paper.bib
@@ -19,7 +19,7 @@
 
 @article{cpp:14,
   author = {ISO/IEC},
-  title = {International standard ISO/IEC 14882:2014(E) – Programming language C++},
+  title = {International standard ISO/IEC 14882:2014(E) – Programming language {C}++},
   year = {2014},
   journal = {Geneva, Switzerland: International Organization for Standardization},
 }
