### 2022-03-11 #3965 @jbytecode
Dear @matthieuaussal

I have made some corrections and additions in the paper and the bibliography.

Please apply the pull request if you are agreed with them.
