### 2022-03-11 #3965 @matthieuaussal
Dear @jbytecode <https://github.com/jbytecode>,

It’s done, thanks for the corrections.

> Le 11 mars 2022 à 21:07, Mehmet Hakan Satman ***@***.***> a écrit :
>
>
> Dear @matthieuaussal <https://github.com/matthieuaussal>
> I have made some corrections and additions in the paper and the bibliography.
>
> Please apply the pull request if you are agreed with them.
>
