### 2022-03-12 #3965 @jbytecode
Dear @matthieuaussal

- [ ] I sent another pull request with a small change (macos -> macOS). If you agree with it, please apply.

- [ ] Please re-check the whole paper by means of both language and format.

- [ ] make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the **correct metadata** (**title and author list**), or edit these if that is not the case.

Thank you in advance.
