+++ paper/paper.bib
@@ -15,7 +15,7 @@
       PUBLISHER = {Society for Industrial and Applied Mathematics},
       YEAR = {1999},
       ADDRESS = {Philadelphia, PA},
-      ISBN = {0-89871-447-8 (paperback)} }
+      ISBN = {0-89871-447-8} }
 
 @article{cpp:14,
   author = {ISO/IEC},
