### 2022-03-15 #3965 @xuanxu
The error is caused by an invalid ISBN value, this PR should fix it: https://github.com/leprojetcastor/castor/pull/5
