### 2022-03-15 #3965 @jbytecode
Thank you @mkitti and @pitsianis for consuming your valuable time for reviewing this submission, and thank you @matthieuaussal  for sending your work and submission to JOSS. We have just completed the review and I recommended an acceptance. The Editor in Chief will take the final decision. Hope we will work in later projects and submissions!
