### 2022-03-16 #3965 @matthieuaussal
> Hi @matthieuaussal! I'll take over from here to wrap up.
>
> Can you update the metadata for your Zenodo archive so the title and author list match your JOSS paper?

Done, is it ok now?
