+++ paper/paper.bib
@@ -1,25 +1,26 @@
 @article{bezanson2012julia,
-  title={Julia: A fast dynamic language for technical computing},
-  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
+  title={{Julia:} {A} fast dynamic language for technical computing},
+  author={{Bezanson}, Jeff and {Karpinski}, Stefan and {Shah}, Viral B. and {Edelman}, Alan},
   journal={arXiv preprint arXiv:1209.5145},
   year={2012}
 } 
 
 @BOOK{blaslapack:00,
-      AUTHOR = {Anderson, E. and Bai, Z. and Bischof, C. and
-                Blackford, S. and Demmel, J. and Dongarra, J. and
-                Du Croz, J. and Greenbaum, A. and Hammarling, S. and
-                McKenney, A. and Sorensen, D.},
-      TITLE = {{LAPACK} Users' guide},
+      AUTHOR = {{Anderson}, E. and {Bai}, Z. and {Bischof}, C. and
+                {Blackford}, S. and {Demmel}, J. and {Dongarra}, J. and
+                {Du Croz}, J. and {Greenbaum}, A. and {Hammarling}, S. and
+                {McKenney}, A. and {Sorensen}, D.},
+      TITLE = {{LAPACK} {Users'} guide},
       EDITION = {Third},
-      PUBLISHER = {Society for Industrial and Applied Mathematics},
+      PUBLISHER = {{Society for Industrial and Applied Mathematics}},
       YEAR = {1999},
       ADDRESS = {Philadelphia, PA},
-      ISBN = {0-89871-447-8} }
+      ISBN = {0-89871-447-8} 
+}
 
 @article{cpp:14,
   author = {ISO/IEC},
-  title = {International standard ISO/IEC 14882:2014(E) – Programming language {C}++},
+  title = {International standard ISO/IEC 14882:2014(E) – {Programming} language {C}++},
   year = {2014},
   journal = {Geneva, Switzerland: International Organization for Standardization},
 }
@@ -27,7 +28,7 @@
 @article{hackbush:1999,
   author = {{Hackbusch}, W.},
   journal = {Computing, 62(2), 89–108},
-  title = {{A sparse matrix arithmetic based on H-matrices. Part 1: Introduction to  H-matrices}},
+  title = {{A sparse matrix arithmetic based on H-matrices. Part 1: Introduction to H-matrices}},
   year = {1999}
 }
 
@@ -49,27 +50,27 @@ address = {Natick, Massachusetts}
 @article{rjasanow:2002,
  author = {{Rjasanow}, S.},
  title = {Adaptive cross approximation of dense matrices},
- journal = {Int. Association Boundary Element Methods Conf., IABEM (pp. 28-30)},
+ journal = {{Int. Association Boundary Element Methods Conf., IABEM (pp. 28-30)}},
  year = 2002
 }
 
 @article{symare:2013,
- author = {Jin, C. T. and Guillon, P. and Epain, N. and Zolfaghari, R. and Van Schaik, A. and Tew, A. I. and  Thorpe, J.},
- title = {Creating the Sydney York morphological and acoustic recordings of ears database},
+ author = {{Jin}, C. T. and {Guillon}, P. and {Epain}, N. and {Zolfaghari}, R. and {van Schaik}, A. and {Tew}, A. I. and {Thorpe}, J.},
+ title = {{Creating the Sydney York morphological and acoustic recordings of ears database}},
  journal = {IEEE Transactions on Multimedia, 16(1), 37-46},
[…]
