### 2022-03-16 #3965 @matthieuaussal
> Le 16 mars 2022 à 18:02, Kristen Thyng ***@***.***> a écrit :
>
>
> @matthieuaussal <https://github.com/matthieuaussal> your paper looks good, except please go through the capitalization in the references. For example, Jin et al has uncapitalized proper nouns. You can preserve capitalization in your .bib file by placing {} around words or phrases.
>
