### 2022-03-16 #3965 @matthieuaussal
> @matthieuaussal your paper looks good, except please go through the capitalization in the references. For example, Jin et al has uncapitalized proper nouns. You can preserve capitalization in your .bib file by placing {} around words or phrases.

I think it's done, but I can't regenerate pdf to verify
