### 2022-10-11 #3965 @matthieuaussal
> Le 10 mars 2022 à 19:04, Matthieu Aussal ***@***.***> a écrit :
>
> @editorialbot generate pdf
>
>
>> Le 7 mars 2022 à 17:26, Mehmet Hakan Satman ***@***.*** ***@***.***>> a écrit :
>>
>>
>> Dear @matthieuaussal <https://github.com/matthieuaussal>
>> It seems our reviewers have just completed their major tasks. Please consider applying the latest minor suggestions, then I am going to start my editorial stuff.
>>
>> Thank you in advance.
>>
>> —
>> Reply to this email directly, view it on GitHub <https://github.com/openjournals/joss-reviews/issues/3965#issuecomment-1060878219>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AHFHRGE27XU5QR5PYNASYGLU6YU4JANCNFSM5JDT4MGQ>.
>> Triage notifications on the go with GitHub Mobile for iOS <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>> You are receiving this because you were mentioned.
>>
>
