### 2022-10-11 #3965 @arfon
@matthieuaussal – curious what you're looking to accomplish here? Once a paper has been accepted it's not possible to generate PDF previews again like this sorry.
